Approximations

Approximations

This section of the manual describes the Cartesian decomposition algorithms and the approximation of high-dimensional convex sets using projections.

Approximations

Cartesian Decomposition

Box Approximations

Metric properties of sets

Iterative refinement

Cartesian Decomposition

LazySets.Approximations.decompose — Function.

decompose(X)

Compute an overapproximation of the projections of the given set over each two-dimensional subspace using box directions.

Input

X – lazy set

Output

A CartesianProductArray corresponding to the cartesian product of 2x2 polygons.

source

decompose(X, ɛi)

Compute an overapproximation of the projections of the given set over each two-dimensional subspace with a certified error bound.

Input

X – lazy set

ɛi – array with the error bound for each projection (different error bounds can be passed to different blocks)

Output

A CartesianProductArray corresponding to the cartesian product of 2x2 polygons.

Algorithm

This algorithm assumes a decomposition into two-dimensional subspaces only, i.e. partitions of the form $[2, 2, ..., 2]$. In particular if X is a CartesianProductArray, no check is performed to verify that assumption.

It proceeds as follows:

Project the set X into each partition, with MX, where M is the identity matrix in the block coordinates and zero otherwise.

Overapproximate the set with a given error bound, ɛi[i], for $i = 1, …, b$,

Return the result as an array of support functions.

source

decompose(X, ɛ)

Compute an overapproximation of the projections of the given set over each two-dimensional subspace with a certified error bound.

This function is a particular case of decompose(X, ɛi), where the same error bound for each block is assumed.

Input

X – lazy set

ɛ – error bound

Output

A CartesianProductArray corresponding to the cartesian product of 2x2 polygons.

source

LazySets.Approximations.overapproximate — Function.

overapproximate(X)

Return an approximation of the given 2D set as a box-shaped polygon.

Input

X – lazy set, assumed to be two-dimensional

Output

A polygon in constraint representation.

source

overapproximate(X, ɛ)

Return an ɛ-close approximation of the given 2D set (in terms of Hausdorff distance) as a polygon.

Input

X – lazy set, assumed to be two-dimensional

ɛ – the error bound

Output

A polygon in constraint representation.

source

Box Approximations

LazySets.Approximations.ballinf_approximation — Function.

ballinf_approximation(X)

Overapproximation of a set by a ball in the infinity norm.

Input

X – a lazy set

Output

H – a ball in the infinity norm which tightly contains the given set

Algorithm

The center and radius of the box are obtained by evaluating the support function of the given set along the canonical directions.

source

LazySets.Approximations.box_approximation — Function.

box_approximation(X)

Overapproximate a set by a box (hyperrectangle).

Input

X – a lazy set

Output

H – a (tight) hyperrectangle

Algorithm

The center of the hyperrectangle is obtained by averaring the support function of the given set in the canonical directions, and the lengths of the sides can be recovered from the distance among support functions in the same directions.

source

LazySets.Approximations.box_approximation_symmetric — Function.

box_approximation_symmetric(X)

Overapproximation of a set by a hyperrectangle which contains the origin.

Input

X – a lazy set

Ouptut

H – a symmetric interval around the origin which tightly contains the given set

Algorithm

The center of the box is the origin, and the radius is obtained by computing the maximum value of the support function evaluated at the canonical directions.

source

LazySets.Approximations.box_approximation_helper — Function.

box_approximation_helper(X)

Common code of box_approximation and box_approximation_symmetric.

Input

X – a lazy set

Output

H – a (tight) hyperrectangle

Algorithm

The center of the hyperrectangle is obtained by averaring the support function the given set in the canonical directions, and the lengths of the sides can be recovered from the distance among support functions in the same directions.

source

Metric properties of sets

Base.LinAlg.norm — Function.

norm(X::LazySet, [p])

Return the norm of a LazySet. It is the norm of the enclosing ball (of the given norm) of minimal volume.

Input

X – a lazy set

p – (optional, default: Inf) norm

Output

A real number representing the norm.

source

LazySets.radius — Function.

radius(X::LazySet, [p])

Return the radius of a LazySet. It is the radius of the enclosing ball (of the given norm) of minimal volume with the same center.

Input

X – lazy set

p – (optional, default: Inf) norm

Output

A real number representing the radius.

source

LazySets.diameter — Function.

diameter(X::LazySet, [p])

Return the diameter of a LazySet. It is the maximum distance between any two elements of the set, or, equivalently, the diameter of the enclosing ball (of the given norm) of minimal volume with the same center.

Input

X – lazy set

p – (optional, default: Inf) norm

Output

A real number representing the diameter.

source

Iterative refinement

LazySets.Approximations.approximate — Function.

approximate(X, ɛ)

Return an ɛ-close approximation of the given 2D set (in terms of Hausdorff distance) as an inner and an outer approximation composed by sorted local Approximation2D.

Input

X – a 2D set defined by its support function

ɛ – the error bound

source