Overapproximation

Overapproximation

LazySets.Approximations.taylor_expmap_truncation — Function

taylor_expmap_truncation(A::MatrixZonotope, P::S, k::Int)
    where {S<:Union{SparsePolynomialZonotope,AbstractZonotope}}

Compute the k-th order truncated Taylor expansion of the exponential map of a matrix zonotope

Input

A – a matrix zonotope

P – a (potentially polynomial) zonotopic set

k – the order of the Taylor expansion

Output

A (polynomial) zonotopic set representing the k-th order truncated Taylor expansion.

Algorithm

This function computes the approximation:

\[\displaystyle\boxplus_{i=0}^k \frac{\mathcal{A}^i }{i!} X\]

taylor_expmap_truncation(MZP::MatrixZonotopeProduct, P::S, k::Int)
    where {S<:Union{SparsePolynomialZonotope,AbstractZonotope}}

Compute the k-th order truncated Taylor expansion of the exponential map of a matrix zonotope product

Input

MZP – a matrix zonotope product

P – a (potentially polynomial) zonotopic set

k – the order of the Taylor expansion

Output

A (potentially polynomial) zonotopic set representing the k-th order truncated Taylor expansion.

Algorithm

This function computes the approximation:

\[\displaystyle\boxplus_{i=0}^k \frac{\mathcal{A}^i \mathcal{B}^i}{i!} P\]

LazySets.Approximations.taylor_expmap_remainder — Function

taylor_expmap_remainder(Z::AbstractZonotope{N}, matnorm::Real, Int) where {N}

Overapproximate the Lagrange remainder term of the k-th order truncated Taylor expansion of the exponential map of a matrix zonotope applied to a zonotopic set.

Input

P – a zonotopic set

matnorm – an upper bound on the norm of the matrix zonotope

k – the order of the Taylor expansion

Output

A zonotope over-approximating the remainder term of the Taylor expansion.

LazySets.Approximations._compute_inner_powers — Function

_compute_inner_powers(B::MatrixZonotope, P::S,
                           k::Int) where {S<:Union{SparsePolynomialZonotope,
                                                   AbstractZonotope}}

Compute the first k powers of the matrix zonotope B applied to the set P.

This function returns a vector of overapproximated sets of the form:

\[ (1/i!) * \mathcal{B}^i * P\]

for i = 0 to k.

LazySets.Approximations._compute_outer_powers — Function

_compute_outer_powers(A::MatrixZonotope, in_powers::Vector{S},
                           k::Int) where {S<:Union{SparsePolynomialZonotope,
                                                   AbstractZonotope}}

Apply A repeatedly to each element in in_powers, approximating:

\[A^i * (B^i * P)\]

for i = 0 to k.