Polynomial zonotope sets (AbstractPolynomialZonotope)

Implementations

Polynomial zonotope sets (AbstractPolynomialZonotope)

LazySets.AbstractPolynomialZonotope — Type

AbstractPolynomialZonotope{N} <: LazySet{N}

Abstract type for polynomial zonotope sets.

Notes

See DensePolynomialZonotope for a standard implementation of this interface.

Polynomial zonotopes are in general non-convex. They are always bounded.

Every concrete AbstractPolynomialZonotope must define the following functions:

center(::AbstractPolynomialZonotope) – return the center

polynomial_order(::AbstractPolynomialZonotope) – return the polynomial order

ngens_dep – return the number of dependent generators

ngens_indep – return the number of independent generators

The subtypes of AbstractPolynomialZonotope (including abstract interfaces):

julia> subtypes(AbstractPolynomialZonotope)
2-element Vector{Any}:
 AbstractSparsePolynomialZonotope
 DensePolynomialZonotope

This interface requires to implement the following functions:

LazySets.API.center — Method

center(X::LazySet)

Compute the center of a centrally symmetric set.

Input

X – centrally symmetric set

Output

A vector with the center, or midpoint, of X.

LazySets.ngens_dep — Method

ngens_dep(P::AbstractPolynomialZonotope)

Determine the number of dependent generators of a polynomial zonotope.

Input

P – polynomial zonotope

Output

A nonnegative integer representing the number of dependent generators.

LazySets.ngens_indep — Method

ngens_indep(P::AbstractPolynomialZonotope)

Determine the number of independent generators of a polynomial zonotope.

Input

P – polynomial zonotope

Output

A nonnegative integer representing the number of independent generators.

LazySets.polynomial_order — Method

polynomial_order(P::AbstractPolynomialZonotope)

Determine the polynomial order of a polynomial zonotope.

Input

P – polynomial zonotope

Output

A nonnegative integer representing the polynomial order.

Notes

The polynomial order is the maximum sum of all monomials' parameter exponents.

This interface defines the following functions:

LazySets.ngens — Method

ngens(P::AbstractPolynomialZonotope)

Return the number of generators of a polynomial zonotope.

Input

P – polynomial zonotope

Output

The total number of generators of P.

LazySets.order — Method

order(P::AbstractPolynomialZonotope)

Return the order of a polynomial zonotope.

Input

P – polynomial zonotope

Output

The order, defined as the quotient between the total number of generators and the ambient dimension, as a Rational number.

LazySets.API.convex_hull — Method

Extended help

convex_hull(PZ1::AbstractPolynomialZonotope, PZ2::AbstractPolynomialZonotope)

Output

The tightest convex polynomial zonotope containing P1 and P2.

Undocumented implementations:

center

convex_hull

dim

isboundedtype

isempty

isuniversal

Inherited from LazySet:

an_element

area

chebyshev_center_radius

complement

concretize

constraints

copy(::Type{LazySet})

diameter

eltype

eltype

extrema

extrema

high

high

isbounded

isoperation

ispolyhedral

ispolyhedraltype

ispolytopic

low

low

norm

polyhedron

radius

rationalize

rectify

reflect

singleton_list

tosimplehrep

triangulate

triangulate_faces

vertices

affine_map

exponential_map

is_interior_point

linear_map

project

sample

scale

ρ

translate

cartesian_product

exact_sum

isapprox

isdisjoint

==

isequivalent

⊂

⊆

minkowski_difference

minkowski_sum

Implementations

Dense polynomial zonotope (DensePolynomialZonotope)