Bloating

LazySets.Bloating — Type

Bloating{N, S<:LazySet{N}} <:LazySet{N}

Type that represents a uniform expansion of a convex set in a given norm (also known as bloating).

Fields

X – convex set

ε – (positive) bloating factor

p – $p$-norm ($≥ 1$; default: $2$)

Notes

Bloating(X, ε, p) is equivalent to the Minkowski sum of X and a ball in the p-norm of radius ε centered in the origin O (i.e., X ⊕ Ballp(p, O, ε)).

LazySets.dim — Method

dim(B::Bloating)

Return the dimension of a bloated convex set.

Input

B – bloated convex set

Output

The ambient dimension of the bloated set.

LazySets.σ — Method

σ(d::AbstractVector, B::Bloating)

Return the support vector of a bloated convex set in a given direction.

Input

d – direction

B – bloated convex set

Output

The support vector of the bloated set in the given direction.

LazySets.ρ — Method

ρ(d::AbstractVector, B::Bloating)

Return the support function of a bloated convex set in a given direction.

Input

d – direction

B – bloated convex set

Output

The support function of the bloated set in the given direction.

LazySets.isbounded — Method

isbounded(B::Bloating)

Determine whether a bloated convex set is bounded.

Input

B – bloated convex set

Output

true iff the wrapped set is bounded.

Base.isempty — Method

isempty(B::Bloating)

Determine whether a bloated convex set is empty.

Input

B – bloated convex set

Output

true iff the wrapped set is empty.

LazySets.an_element — Method

an_element(B::Bloating)

Return some element of a bloated convex set.

Input

B – bloated convex set

Output

An element in the bloated convex set.

Algorithm

The implementation returns the result of an_element for the wrapped set.

LazySets.constraints_list — Method

constraints_list(B::Bloating)

Return the list of constraints of a bloated set.

Input

B – bloated set

Output

The list of constraints of the bloated set.

Notes

The constraints list is only available for bloating in the p-norm for $p = 1$ or $p = ∞$ and if constraints_list is available for the unbloated set.

Algorithm

We compute a concrete set representation via minkowski_sum and call constraints_list on the result.

Inherited from LazySet:

singleton_list