Common Set Operations

Common Set Operations

This section of the manual describes the basic symbolic types describing operations between sets.

Common Set Operations

Minkowski Sum

Binary Minkowski Sum

$n$-ary Minkowski Sum

Cartesian Product

Binary Cartesian Product

$n$-ary Cartesian Product

Maps

Linear Map

Exponential Map

Convex Hull

Convex Hull Algorithms

Minkowski Sum

Binary Minkowski Sum

LazySets.MinkowskiSum — Type.

MinkowskiSum{T1<:LazySet, T2<:LazySet} <: LazySet

Type that represents the Minkowski sum of two convex sets.

Fields

X – first convex set

Y – second convex set

source

LazySets.dim — Method.

dim(ms::MinkowskiSum)::Int

Return the dimension of a Minkowski sum.

Input

ms – Minkowski sum

Output

The ambient dimension of the Minkowski sum.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, ms::MinkowskiSum)::AbstractVector{<:Real}

Return the support vector of a Minkowski sum.

Input

d – direction

ms – Minkowski sum

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the summand sets.

source

Base.:+ — Method.

+(X::LazySet, Y::LazySet)

Return the Minkowski sum of two convex sets.

Input

X – convex set

Y – convex set

Output

The Minkowski sum of the two convex sets.

source

$n$-ary Minkowski Sum

LazySets.MinkowskiSumArray — Type.

MinkowskiSumArray{T<:LazySet} <: LazySet

Type that represents the Minkowski sum of a finite number of convex sets.

Fields

sfarray – array of convex sets

Notes

This type assumes that the dimensions of all elements match.

MinkowskiSumArray(sfarray::Vector{<:LazySet}) – default constructor

MinkowskiSumArray() – constructor for an empty sum

MinkowskiSumArray(n::Integer) – constructor for an empty sum with size hint

source

LazySets.dim — Method.

dim(msa::MinkowskiSumArray)::Int

Return the dimension of a Minkowski sum of a finite number of sets.

Input

msa – Minkowski sum array

Output

The ambient dimension of the Minkowski sum of a finite number of sets.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, msa::MinkowskiSumArray)::Vector{<:Real}

Return the support vector of a Minkowski sum of a finite number of sets in a given direction.

Input

d – direction

msa – Minkowski sum array

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the summand sets.

source

Base.:+ — Method.

+(msa::MinkowskiSumArray, S::LazySet)::MinkowskiSumArray

Add a convex set to a Minkowski sum of a finite number of convex sets from the right.

Input

msa – Minkowski sum array (is modified)

S – convex set

Output

The modified Minkowski sum of a finite number of convex sets.

source

Base.:+ — Method.

+(S::LazySet, msa::MinkowskiSumArray)::MinkowskiSumArray

Add a convex set to a Minkowski sum of a finite number of convex sets from the left.

Input

S – convex set

msa – Minkowski sum array (is modified)

Output

The modified Minkowski sum of a finite number of convex sets.

source

Base.:+ — Method.

+(msa1::MinkowskiSumArray, msa2::MinkowskiSumArray)::MinkowskiSumArray

Add the elements of a finite Minkowski sum of convex sets to another finite Minkowski sum.

Input

msa1 – first Minkowski sum array (is modified)

msa2 – second Minkowski sum array

Output

The modified first Minkowski sum of a finite number of convex sets.

source

Base.:+ — Method.

+(msa, vs)

Returns the original array because addition with a void set is a no-op.

Input

msa – Minkowski sum array

vs – void set

source

Cartesian Product

Binary Cartesian Product

LazySets.CartesianProduct — Type.

CartesianProduct{S1<:LazySet,S2<:LazySet} <: LazySet

Type that represents a Cartesian product of two convex sets.

Fields

X – first convex set

Y – second convex set

Notes

The Cartesian product of three elements is obtained recursively. See also CartesianProductArray for an implementation of a Cartesian product of many sets without recursion, instead using an array.

CartesianProduct{S1<:LazySet,S2<:LazySet} – default constructor

CartesianProduct(Xarr::Vector{S}) where {S<:LazySet} – constructor from an array of convex sets

source

LazySets.dim — Method.

dim(cp::CartesianProduct)::Int

Return the dimension of a Cartesian product.

Input

cp – Cartesian product

Output

The ambient dimension of the Cartesian product.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, cp::CartesianProduct)::AbstractVector{<:Real}

Return the support vector of a Cartesian product.

Input

d – direction

cp – Cartesian product

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the product sets.

source

Base.:* — Method.

    *(X::LazySet, Y::LazySet)::CartesianProduct

Return the Cartesian product of two convex sets.

Input

X – convex set

Y – convex set

Output

The Cartesian product of the two convex sets.

source

LazySets.is_contained — Method.

is_contained(x::AbstractVector{<:Real}, cp::CartesianProduct)::Bool

Return whether a given vector is contained in a Cartesian product set.

Input

x – vector

cp – Cartesian product

Output

Return true iff $x ∈ cp$.

source

$n$-ary Cartesian Product

LazySets.CartesianProductArray — Type.

CartesianProductArray{S<:LazySet} <: LazySet

Type that represents the Cartesian product of a finite number of convex sets.

Fields

sfarray – array of sets

Notes

CartesianProductArray(sfarray::Vector{<:LazySet}) – default constructor

CartesianProductArray() – constructor for an empty Cartesian product

CartesianProductArray(n::Integer) – constructor for an empty Cartesian product with size hint

source

LazySets.dim — Method.

dim(cpa::CartesianProductArray)::Int

Return the dimension of a Cartesian product of a finite number of convex sets.

Input

cpa – Cartesian product array

Output

The ambient dimension of the Cartesian product of a finite number of convex sets.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, cpa::CartesianProductArray)::AbstractVector{<:Real}

Support vector of a Cartesian product.

Input

d – direction

cpa – Cartesian product array

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the product sets.

source

Base.:* — Method.

    *(cpa::CartesianProductArray, S::LazySet)::CartesianProductArray

Multiply a convex set to a Cartesian product of a finite number of convex sets from the right.

Input

cpa – Cartesian product array (is modified)

S – convex set

Output

The modified Cartesian product of a finite number of convex sets.

source

Base.:* — Method.

    *(S::LazySet, cpa::CartesianProductArray)::CartesianProductArray

Multiply a convex set to a Cartesian product of a finite number of convex sets from the left.

Input

S – convex set

cpa – Cartesian product array (is modified)

Output

The modified Cartesian product of a finite number of convex sets.

source

Base.:* — Method.

    *(cpa1::CartesianProductArray, cpa2::CartesianProductArray)::CartesianProductArray

Multiply a finite Cartesian product of convex sets to another finite Cartesian product.

Input

cpa1 – first Cartesian product array (is modified)

cpa2 – second Cartesian product array

Output

The modified first Cartesian product.

source

LazySets.is_contained — Method.

is_contained(x::AbstractVector{<:Real}, cpa::CartesianProductArray)::Bool

Return whether a given vector is contained in a Cartesian product of a finite number of sets.

Input

x – vector

cpa – Cartesian product array

Output

Return true iff $x ∈ cpa$.

source

Maps

Linear Map

LazySets.LinearMap — Type.

LinearMap{S<:LazySet, N<:Real} <: LazySet

Type that represents a linear transformation $M⋅S$ of a convex set $S$.

Fields

M – matrix/linear map

sf – convex set

source

LazySets.dim — Method.

dim(lm::LinearMap)::Int

Return the dimension of a linear map.

Input

lm – linear map

Output

The ambient dimension of the linear map.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, lm::LinearMap)::AbstractVector{<:Real}

Return the support vector of the linear map.

Input

d – direction

lm – linear map

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the wrapped set.

Notes

If $L = M⋅S$, where $M$ is a matrix and $S$ is a convex set, it follows that $σ(d, L) = M⋅σ(M^T d, S)$ for any direction $d$.

source

Base.:* — Method.

    *(M::AbstractMatrix{<:Real}, S::LazySet)

Return the linear map of a convex set.

Input

M – matrix/linear map

S – convex set

Output

The linear map of the convex set.

source

Base.:* — Method.

    *(a::Real, S::LazySet)::LinearMap

Return a linear map of a convex set by a scalar value.

Input

a – real scalar

S – convex set

Output

The linear map of the convex set.

source

Exponential Map

LazySets.ExponentialMap — Type.

ExponentialMap{S<:LazySet} <: LazySet

Type that represents the action of an exponential map on a convex set.

Fields

spmexp – sparse matrix exponential

X – convex set

source

LazySets.dim — Method.

dim(em::ExponentialMap)::Int

Return the dimension of an exponential map.

Input

em – an ExponentialMap

Output

The ambient dimension of the exponential map.

source

LazySets.σ — Method.

σ(d::AbstractVector{Float64}, em::ExponentialMap)::AbstractVector{Float64}

Return the support vector of the exponential map.

Input

d – direction

em – exponential map

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the wrapped set.

Notes

If $E = \exp(M)⋅S$, where $M$ is a matrix and $S$ is a convex set, it follows that $σ(d, E) = \exp(M)⋅σ(\exp(M)^T d, S)$ for any direction $d$.

source

LazySets.ExponentialProjectionMap — Type.

ExponentialProjectionMap{S<:LazySet} <: LazySet

Type that represents the application of a projection of a sparse matrix exponential to a convex set.

Fields

spmexp – projection of a sparse matrix exponential

X – convex set

source

LazySets.dim — Method.

dim(eprojmap::ExponentialProjectionMap)::Int

Return the dimension of a projection of an exponential map.

Input

eprojmap – projection of an exponential map

Output

The ambient dimension of the projection of an exponential map.

source

LazySets.σ — Method.

σ(d::AbstractVector{Float64}, eprojmap::ExponentialProjectionMap)::AbstractVector{Float64}

Return the support vector of a projection of an exponential map.

Input

d – direction

eprojmap – projection of an exponential map

Output

The support vector in the given direction. If the direction has norm zero, the result depends on the wrapped set.

Notes

If $S = (L⋅M⋅R)⋅X$, where $L$ and $R$ are matrices, $M$ is a matrix exponential, and $X$ is a set, it follows that $σ(d, S) = L⋅M⋅R⋅σ(R^T⋅M^T⋅L^T⋅d, X)$ for any direction $d$.

source

LazySets.SparseMatrixExp — Type.

SparseMatrixExp{N<:Real}

Type that represents the matrix exponential, $\exp(M)$, of a sparse matrix.

Fields

M – sparse matrix

Notes

This type is provided for use with very large and very sparse matrices. The evaluation of the exponential matrix action over vectors relies on the Expokit package.

source

Base.:* — Method.

    *(spmexp::SparseMatrixExp, X::LazySet)

Return the exponential map of a convex set from a sparse matrix exponential.

Input

spmexp – sparse matrix exponential

X – convex set

Output

The exponential map of the convex set.

source

LazySets.ProjectionSparseMatrixExp — Type.

ProjectionSparseMatrixExp{N<:Real}

Type that represents the projection of a sparse matrix exponential, i.e., $L⋅\exp(M)⋅R$ for a given sparse matrix $M$.

Fields

L – left multiplication matrix

E – sparse matrix exponential

R – right multiplication matrix

source

Base.:* — Method.

    *(projspmexp::ProjectionSparseMatrixExp, X::LazySet)::ExponentialProjectionMap

Return the application of a projection of a sparse matrix exponential to a convex set.

Input

projspmexp – projection of a sparse matrix exponential

X – convex set

Output

The application of the projection of a sparse matrix exponential to the convex set.

source

Convex Hull

LazySets.ConvexHull — Type.

ConvexHull{S1<:LazySet, S2<:LazySet} <: LazySet

Type that represents the convex hull of the union of two convex sets.

Fields

X – convex set

Y – convex set

source

LazySets.CH — Type.

CH

Alias for ConvexHull.

source

LazySets.dim — Method.

dim(ch::ConvexHull)::Int

Return the dimension of a convex hull of two convex sets.

Input

ch – convex hull of two convex sets

Output

The ambient dimension of the convex hull of two convex sets.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, ch::ConvexHull)::AbstractVector{<:Real}

Return the support vector of a convex hull of two convex sets in a given direction.

Input

d – direction

ch – convex hull of two convex sets

source

Convex Hull Algorithms

LazySets.convex_hull — Function.

convex_hull(points; [algorithm]::String="monotone_chain")

Compute the convex hull of points in the plane.

Input

points – list of 2D vectors

algorithm – (optional, default: "monotone_chain") the convex hull algorithm, valid options are:

"monotone_chain"

Output

The convex hull as a list of 2D vectors with the coordinates of the points.

Examples

Compute the convex hull of a random set of points:

julia> points = [randn(2) for i in 1:30]; # 30 random points in 2D

julia> hull = convex_hull(points);

julia> typeof(hull)
Array{Array{Float64,1},1}

Plot both the random points and the computed convex hull polygon:

julia> using Plots;

julia> plot([Tuple(pi) for pi in points], seriestype=:scatter);

julia> plot!(VPolygon(hull), alpha=0.2);

source

LazySets.convex_hull! — Function.

convex_hull!(points; algorithm::String="monotone_chain")

Compute the convex hull of points in the plane, in-place.

Input

points – list of 2D vectors (is modified)

algorithm – (optional, default: "monotone_chain") the convex hull algorithm, valid options are:

"monotone_chain"

Notes

See the non-modifying version convex_hull for more details.

source

LazySets.right_turn — Function.

right_turn(O, A, B)

Determine if the acute angle defined by the three points O, A, B in the plane is a right turn (counter-clockwise) with respect to the center O.

Input

O – center point

A – one point

B – another point

Algorithm

The cross product is used to determine the sense of rotation. If the result is 0, the points are collinear; if it is positive, the three points constitute a positive angle of rotation around O from A to B; otherwise they constitute a negative angle.

source

LazySets.monotone_chain! — Function.

monotone_chain!(points::Vector{S}) where {S<:AbstractVector{N}} where {N<:Real}

Compute the convex hull of points in the plane using Andrew's monotone chain method.

Input

points – list of 2D vectors; is sorted in-place inside this function

Output

List of vectors containing the 2D coordinates of the corner points of the convex hull.

Notes

For large sets of points, it is convenient to use static vectors to get maximum performance. For information on how to convert usual vectors into static vectors, see the type SVector provided by the StaticArrays package.

Algorithm

This function implements Andrew's monotone chain convex hull algorithm to construct the convex hull of a set of $n$ points in the plane in $O(n \log n)$ time. For further details see Monotone chain

source