Common Set Representations

Common Set Representations

This section of the manual describes the basic set representation types.

Common Set Representations

Abstract support function and support vector

Balls

Euclidean norm ball

Infinity norm ball

Polygons

Constraint representation

Optimized constraint representation

Vertex representation

Lines and linear constraints

Hyperrectangles

VoidSets

Singletons

Zonotopes

Abstract support function and support vector

LazySets — Module.

Main module for LazySets.jl – a Julia package for calculus with convex sets.

source

LazySets.LazySet — Type.

LazySet

Abstract type for a lazy set.

Notes

Every concrete LazySet must define a function σ(d, X), representing the support vector of X in a given direction d, and dim, the ambient dimension of the set X.

LazySet types should be parameterized with a type N, typically N<:Real, to support computations with different numeric types.

source

LazySets.ρ — Function.

ρ(d::AbstractVector{N}, S::LazySet)::N where {N<:Real}

Evaluate the support function of a set in a given direction.

Input

d – direction

S – convex set

Output

The support function of the set S for the direction d.

source

LazySets.support_function — Function.

support_function

Alias for the support function ρ.

source

LazySets.support_vector — Function.

support_vector

Alias for the support vector σ.

source

Balls

Euclidean norm ball

LazySets.Ball2 — Type.

Ball2{N<:Real} <: LazySet

Type that represents a ball in the 2-norm.

Fields

center – center of the ball as a real vector

radius – radius of the ball as a real scalar ($≥ 0$)

Notes

Mathematically, a ball in the 2-norm is defined as the set

\[\mathcal{B}_2^n(c, r) = \{ x ∈ \mathbb{R}^n : ‖ x - c ‖_2 ≤ r \},\]

where $c ∈ \mathbb{R}^n$ is its center and $r ∈ \mathbb{R}_+$ its radius. Here $‖ ⋅ ‖_2$ denotes the Euclidean norm (also known as 2-norm), defined as $‖ x ‖_2 = \left( \sum\limits_{i=1}^n |x_i|^2 \right)^{1/2}$ for any $x ∈ \mathbb{R}^n$.

Examples

Create a five-dimensional ball B in the 2-norm centered at the origin with radius 0.5:

julia> B = Ball2(zeros(5), 0.5)
LazySets.Ball2{Float64}([0.0, 0.0, 0.0, 0.0, 0.0], 0.5)
julia> dim(B)
5

Evaluate B's support vector in the direction $[1,2,3,4,5]$:

julia> σ([1.,2.,3.,4.,5.], B)
5-element Array{Float64,1}:
 0.06742
 0.13484
 0.20226
 0.26968
 0.3371

source

LazySets.dim — Method.

dim(B::Ball2)

Return the dimension of a ball in the 2-norm.

Input

B – ball in the 2-norm

Output

The ambient dimension of the ball.

source

LazySets.σ — Method.

σ(d::AbstractVector{N}, B::Ball2)::AbstractVector{<:AbstractFloat} where {N<:AbstractFloat}

Return the support vector of a 2-norm ball in a given direction.

Input

d – direction

B – ball in the 2-norm

Output

The support vector in the given direction. If the direction has norm zero, the origin is returned.

Notes

Let c and r be the center and radius of a ball B in the 2-norm, respectively. For nonzero direction d, we have σ(d, B) = c + d * (r / ||d||).

This function requires computing the 2-norm of the input direction, which is performed in the given precision of the numeric datatype of both the direction and the set. Exact inputs are not supported.

source

Infinity norm ball

LazySets.BallInf — Type.

BallInf{N<:Real} <: LazySet

Type that represents a ball in the infinity norm.

Fields

center – center of the ball as a real vector

radius – radius of the ball as a real scalar ($≥ 0$)

Notes

Mathematically, a ball in the infinity norm is defined as the set

\[\mathcal{B}_∞^n(c, r) = \{ x ∈ \mathbb{R}^n : ‖ x - c ‖_∞ ≤ r \},\]

where $c ∈ \mathbb{R}^n$ is its center and $r ∈ \mathbb{R}_+$ its radius. Here $‖ ⋅ ‖_∞$ denotes the infinity norm, defined as $‖ x ‖_∞ = \max\limits_{i=1,…,n} \vert x_i \vert$ for any $x ∈ \mathbb{R}^n$.

Examples

Create the two-dimensional unit ball and compute its support function along the positive $x=y$ direction:

julia> B = BallInf(zeros(2), 1.0)
LazySets.BallInf{Float64}([0.0, 0.0], 1.0)
julia> dim(B)
2
julia> ρ([1., 1.], B)
2.0

source

LazySets.dim — Method.

dim(B::BallInf)

Return the dimension of a ball in the infinity norm.

Input

B – ball in the infinity norm

Output

The ambient dimension of the ball.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, B::BallInf)::AbstractVector{<:Real}

Return the support vector of an infinity norm ball in a given direction.

Input

d – direction

B – ball in the infinity norm

Output

The support vector in the given direction. If the direction has norm zero, the vertex with biggest values is returned.

source

LazySets.vertices_list — Method.

vertices_list(B::BallInf{N})::Vector{Vector{N}} where {N<:Real}

Return the list of vertices of a ball in the infinity norm.

Input

B – ball in the infinity norm

Output

A list of vertices.

Notes

For high dimensions, it is preferable to develop a vertex_iterator approach.

Examples

julia> B = BallInf(zeros(2), 0.1)
LazySets.BallInf{Float64}([0.0, 0.0], 0.1)
julia> vertices_list(B)
4-element Array{Array{Float64,1},1}:
 [0.1, 0.1]
 [-0.1, 0.1]
 [0.1, -0.1]
 [-0.1, -0.1]

source

Base.LinAlg.norm — Function.

norm(B::BallInf, [p]::Real=Inf)::Real

Return the norm of a ball in the infinity norm.

Input

B – ball in the infinity norm

p – (optional, default: Inf) norm

Output

A real number representing the norm.

Notes

The norm of an infinity ball is defined as the norm of the enclosing ball, of the given $p$-norm, of minimal volume.

source

LazySets.radius — Function.

radius(B::BallInf, [p]::Real=Inf)

Return the radius of a ball in the infinity norm.

Input

B – ball in the infinity norm

p – (optional, default: Inf) norm

Output

A real number representing the radius.

Notes

The radius is defined as the radius of the enclosing ball of the given $p$-norm of minimal volume with the same center.

source

LazySets.diameter — Function.

diameter(B::BallInf, [p]::Real=Inf)

Return the diameter of a ball in the infinity norm.

Input

B – ball in the infinity norm

p – (optional, default: Inf) norm

Output

A real number representing the diameter.

Notes

The diameter is defined as the maximum distance in the given $p$-norm between any two elements of the set. Equivalently, it is the diameter of the enclosing ball of the given $p$-norm of minimal volume with the same center.

source

Polygons

Constraint representation

LazySets.HPolygon — Type.

HPolygon{N<:Real} <: LazySet

Type that represents a convex polygon in constraint representation whose edges are sorted in counter-clockwise fashion with respect to their normal directions.

Fields

constraints_list – list of linear constraints, sorted by the angle

Notes

The default constructor assumes that the given list of edges is sorted. It does not perform any sorting. Use addconstraint! to iteratively add the edges in a sorted way.

HPolygon(constraints_list::Vector{LinearConstraint{<:Real}}) – default constructor

HPolygon() – constructor with no constraints

source

LazySets.addconstraint! — Method.

addconstraint!(P::HPolygon{N}, constraint::LinearConstraint{N}) where {N<:Real}

Add a linear constraint to a polygon in constraint representation, keeping the constraints sorted by their normal directions.

Input

P – polygon

constraint – linear constraint to add

source

LazySets.dim — Method.

dim(P::HPolygon)

Return the dimension of a polygon.

Input

P – polygon in constraint representation

Output

The ambient dimension of the polygon.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, P::HPolygon{N})::Vector{N} where {N<:Real}

Return the support vector of a polygon in a given direction.

Input

d – direction

P – polygon in constraint representation

Output

The support vector in the given direction. The result is always one of the vertices; in particular, if the direction has norm zero, any vertex is returned.

Algorithm

Comparison of directions is performed using polar angles; see the overload of <= for two-dimensional vectors.

source

LazySets.is_contained — Method.

is_contained(x::AbstractVector{<:Real}, P::HPolygon)::Bool

Return whether a given vector is contained in a polygon.

Input

x – two-dimensional vector

P – polygon in constraint representation

Output

Return true iff $x ∈ P$.

source

LazySets.tovrep — Method.

tovrep(P::HPolygon)::VPolygon

Build a vertex representation of the given polygon.

Input

P – polygon in constraint representation

Output

The same polygon but in vertex representation, a VPolygon.

source

LazySets.vertices_list — Method.

vertices_list(P::HPolygon{N})::Vector{Vector{N}} where {N<:Real}

Return the list of vertices of a polygon in constraint representation.

Input

P – polygon in constraint representation

Output

List of vertices.

source

Optimized constraint representation

LazySets.HPolygonOpt — Type.

HPolygonOpt{N<:Real} <: LazySet

Type that represents a convex polygon in constraint representation whose edges are sorted in counter-clockwise fashion with respect to their normal directions. This is a refined version of HPolygon.

Fields

constraints_list – list of linear constraints

ind – index in the list of constraints to begin the search to evaluate the support function

Notes

This structure is optimized to evaluate the support function/vector with a large sequence of directions that are close to each other. The strategy is to have an index that can be used to warm-start the search for optimal values in the support vector computation.

The default constructor assumes that the given list of edges is sorted. It does not perform any sorting. Use addconstraint! to iteratively add the edges in a sorted way.

HPolygonOpt(constraints_list::Vector{LinearConstraint{<:Real}}, ind::Int64) – default constructor

HPolygonOpt(constraints_list::Vector{LinearConstraint{<:Real}}) – constructor without index

HPolygonOpt(H::HPolygon{<:Real}) – constructor from an HPolygon

source

LazySets.addconstraint! — Method.

addconstraint!(P::HPolygonOpt{N}, constraint::LinearConstraint{N}) where {N<:Real}

Add a linear constraint to an optimized polygon in constraint representation, keeping the constraints sorted by their normal directions.

Input

P – optimized polygon

constraint – linear constraint to add

source

LazySets.dim — Method.

dim(P::HPolygonOpt)

Return the dimension of an optimized polygon.

Input

P – optimized polygon in constraint representation

Output

The ambient dimension of the optimized polygon.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, P::HPolygonOpt{N})::Vector{N} where {N<:Real}

Return the support vector of an optimized polygon in a given direction.

Input

d – direction

P – optimized polygon in constraint representation

Output

The support vector in the given direction. The result is always one of the vertices; in particular, if the direction has norm zero, any vertex is returned.

Algorithm

Comparison of directions is performed using polar angles; see the overload of <= for two-dimensional vectors.

source

LazySets.is_contained — Method.

is_contained(x::AbstractVector{<:Real}, P::HPolygonOpt)::Bool

Return whether a given vector is contained in an optimized polygon.

Input

x – two-dimensional vector

P – optimized polygon in constraint representation

Output

Return true iff $x ∈ P$.

source

LazySets.tovrep — Method.

tovrep(P::HPolygonOpt)::VPolygon

Build a vertex representation of the given optimized polygon.

Input

P – optimized polygon in constraint representation

Output

The same polygon but in vertex representation, a VPolygon.

source

LazySets.vertices_list — Method.

vertices_list(P::HPolygonOpt{N})::Vector{Vector{N}} where {N<:Real}

Return the list of vertices of an optimized polygon in constraint representation.

Input

P – an optimized polygon in constraint representation

Output

List of vertices.

source

Vertex representation

LazySets.VPolygon — Type.

VPolygon{N<:Real} <: LazySet

Type that represents a polygon by its vertices.

Fields

vertices_list – the list of vertices

Notes

The constructor of VPolygon runs a convex hull algorithm, and the given vertices are sorted in counter-clockwise fashion. The constructor flag apply_convex_hull can be used to skip the computation of the convex hull.

VPolygon(vertices_list::Vector{Vector{N}}; apply_convex_hull::Bool=true, algorithm::String="monotone_chain")

source

LazySets.dim — Method.

dim(P::VPolygon)

Return the dimension of a polygon in vertex representation.

Input

P – polygon in vertex representation

Output

The ambient dimension of the polygon.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, P::VPolygon{N})::Vector{N} where {N<:Real}

Return the support vector of a polygon in a given direction.

Input

d – direction

P – polygon in vertex representation

Output

The support vector in the given direction. If the direction has norm zero, the first vertex is returned.

Algorithm

This implementation performs a brute-force search, comparing the projection of each vector along the given direction. It runs in $O(n)$ where $n$ is the number of vertices.

Notes

For arbitrary points without structure this is the best one can do. However, a more efficient approach can be used if the vertices of the polygon have been sorted in counter-clockwise fashion. In that case a binary search algorithm can be used that runs in $O(\log n)$. See issue #40.

source

LazySets.vertices_list — Method.

vertices_list(P::VPolygon{N})::Vector{Vector{N}} where {N<:Real}

Return the list of vertices of a convex polygon in vertex representation.

Input

P – a polygon vertex representation

Output

List of vertices.

source

LazySets.singleton_list — Method.

singleton_list(P::VPolygon{N})::Vector{Singleton{N}} where {N<:Real}

Return the vertices of a convex polygon in vertex representation as a list of singletons.

Input

P – a polygon vertex representation

Output

List containing a singleton for each vertex.

source

Lines and linear constraints

LazySets.LinearConstraint — Type.

LinearConstraint{N<:Real}

Type that represents a linear constraint (a half-space) of the form $a⋅x ≤ b$.

Fields

a – normal direction

b – constraint

Examples

The set $y ≥ 0$ in the plane:

julia> LinearConstraint([0, -1.], 0.)
LazySets.LinearConstraint{Float64}([0.0, -1.0], 0.0)

source

LazySets.Line — Type.

Line{N<:Real}

Type that represents a line in 2D of the form $a⋅x = b$.

Fields

a – normal direction

b – constraint

Examples

The line $y = -x + 1$:

julia> Line([1., 1.], 1.)
LazySets.Line{Float64}([1.0, 1.0], 1.0)

source

LazySets.intersection — Method.

intersection(L1::Line{N}, L2::Line{N})::Vector{N} where {N<:Real}

Return the intersection of two 2D lines.

Input

L1 – first line

L2 – second line

Output

The intersection point, if any. Throws a SingularException if the lines do not intersect.

Examples

The line $y = -x + 1$ intersected with the line $y = x$:

julia> intersection(Line([-1., 1.], 0.), Line([1., 1.], 1.))
2-element Array{Float64,1}:
 0.5
 0.5

source

Hyperrectangles

LazySets.Hyperrectangle — Type.

Hyperrectangle{N<:Real} <: LazySet

Type that represents a hyperrectangle.

A hyperrectangle is the Cartesian product of one-dimensional intervals.

Fields

center – center of the hyperrectangle as a real vector

radius – radius of the ball as a real vector, i.e., half of its width along each coordinate direction

source

LazySets.Hyperrectangle — Method.

Hyperrectangle(;kwargs...)

Construct a hyperrectangle from keyword arguments.

Input

kwargs – keyword arguments; two combinations are allowed:

center, radius – vectors

high, low – vectors (if both center and radius are also defined, those are chosen instead)

Output

A hyperrectangle.

Examples

The following three constructions are equivalent:

julia> c = ones(2);

julia> r = [0.1, 0.2];

julia> l = [0.9, 0.8];

julia> h = [1.1, 1.2];

julia> H1 = Hyperrectangle(c, r)
LazySets.Hyperrectangle{Float64}([1.0, 1.0], [0.1, 0.2])
julia> H2 = Hyperrectangle(center=c, radius=r)
LazySets.Hyperrectangle{Float64}([1.0, 1.0], [0.1, 0.2])
julia> H3 = Hyperrectangle(low=l, high=h)
LazySets.Hyperrectangle{Float64}([1.0, 1.0], [0.1, 0.2])

source

LazySets.dim — Method.

dim(H::Hyperrectangle)

Return the dimension of a hyperrectangle.

Input

H – hyperrectangle

Output

The ambient dimension of the hyperrectangle.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, H::Hyperrectangle)::AbstractVector{<:Real}

Return the support vector of a hyperrectangle in a given direction.

Input

d – direction

H – hyperrectangle

Output

The support vector in the given direction. If the direction has norm zero, the vertex with biggest values is returned.

source

LazySets.vertices_list — Method.

vertices_list(H::Hyperrectangle{N})::Vector{Vector{N}} where {N<:Real}

Return the vertices of a hyperrectangle.

Input

H – hyperrectangle

Output

A list of vertices.

Notes

For high dimensions, it is preferable to develop a vertex_iterator approach.

source

Base.LinAlg.norm — Function.

norm(H::Hyperrectangle, [p]::Real=Inf)

Return the norm of a hyperrectangle.

Input

H – hyperrectangle

p – (optional, default: Inf) norm

Output

A real number representing the norm.

Notes

The norm of a hyperrectangle is defined as the norm of the enclosing ball, of the given $p$-norm, of minimal volume.

source

LazySets.radius — Function.

radius(H::Hyperrectangle, [p]::Real=Inf)

Return the radius of a hyperrectangle.

Input

H – hyperrectangle

p – (optional, default: Inf) norm

Output

A real number representing the radius.

Notes

The radius is defined as the radius of the enclosing ball of the given $p$-norm of minimal volume with the same center.

source

LazySets.diameter — Function.

diameter(H::Hyperrectangle, [p]::Real=Inf)

Return the diameter of a hyperrectangle.

Input

H – hyperrectangle

p – (optional, default: Inf) norm

Output

A real number representing the diameter.

Notes

The diameter is defined as the maximum distance in the given $p$-norm between any two elements of the set. Equivalently, it is the diameter of the enclosing ball of the given $p$-norm of minimal volume with the same center.

source

LazySets.high — Method.

high(H::Hyperrectangle)

Return the higher coordinates of a hyperrectangle.

Input

H – hyperrectangle

Output

A vector with the higher coordinates of the hyperrectangle, one entry per dimension.

source

LazySets.low — Method.

low(H::Hyperrectangle)

Return the lower coordinates of a hyperrectangle.

Input

H – hyperrectangle

Output

A vector with the lower coordinates of the hyperrectangle, one entry per dimension.

source

VoidSets

LazySets.VoidSet — Type.

VoidSet <: LazySet

Type that represents a void (neutral) set with respect to Minkowski sum.

Fields

dim – ambient dimension of the VoidSet

source

Singletons

LazySets.Singleton — Type.

Singleton{N<:Real} <: LazySet

Type that represents a singleton, that is, a set with a unique element.

Fields

element – the only element of the set

source

LazySets.dim — Method.

dim(S::Singleton)

Return the dimension of a singleton.

Input

S – singleton

Output

The ambient dimension of the singleton.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, S::LazySets.Singleton{N})::Vector{N} where {N<:Real}

Return the support vector of a singleton.

Input

d – direction

B – singleton

Output

The support vector, which is the singleton's vector itself, irrespective of the given direction.

source

Zonotopes

LazySets.Zonotope — Type.

Zonotope{N<:Real} <: LazySet

Type that represents a zonotope.

Fields

center – center of the zonotope

generators – matrix; each column is a generator of the zonotope

Notes

Mathematically, a zonotope is defined as the set

\[Z = \left\{ c + \sum_{i=1}^p ξ_i g_i,~~ ξ_i \in [-1, 1]~~ ∀ i = 1,…, p \right\},\]

where $c \in \mathbb{R}^n$ is its center and $\{g_i\}_{i=1}^p$, $g_i \in \mathbb{R}^n$, is the set of generators. This characterization defines a zonotope as the finite Minkowski sum of line elements. Zonotopes can be equivalently described as the image of a unit infinity-norm ball in $\mathbb{R}^n$ by an affine transformation.

Zonotope(center::AbstractVector{N}, generators::AbstractMatrix{N}) where {N<:Real}

Zonotope(center::AbstractVector{N}, generators_list::AbstractVector{T}) where {N<:Real, T<:AbstractVector{N}}

Examples

A two-dimensional zonotope with given center and set of generators:

julia> Z = Zonotope([1.0, 0.0], 0.1*eye(2))
LazySets.Zonotope{Float64}([1.0, 0.0], [0.1 0.0; 0.0 0.1])
julia> dim(Z)
2

Compute its vertices:

julia> vertices_list(Z)
4-element Array{Array{Float64,1},1}:
 [0.9, -0.1]
 [1.1, -0.1]
 [1.1, 0.1]
 [0.9, 0.1]

Evaluate the support vector in a given direction:

julia> σ([1., 1.], Z)
2-element Array{Float64,1}:
 1.1
 0.1

Alternative constructor: A zonotope in two dimensions with three generators:

julia> Z = Zonotope(ones(2), [[1., 0.], [0., 1.], [1., 1.]])
LazySets.Zonotope{Float64}([1.0, 1.0], [1.0 0.0 1.0; 0.0 1.0 1.0])
julia> Z.generators
2×3 Array{Float64,2}:
 1.0  0.0  1.0
 0.0  1.0  1.0

source

LazySets.dim — Method.

dim(Z::Zonotope)

Return the dimension of a zonotope.

Input

Z – zonotope

Output

The ambient dimension of the zonotope.

source

LazySets.σ — Method.

σ(d::AbstractVector{<:Real}, Z::Zonotope)::AbstractVector{<:Real}

Return the support vector of a zonotope in a given direction.

Input

d – direction

Z – zonotope

Output

Support vector in the given direction.

source

LazySets.vertices_list — Method.

vertices_list(Z::Zonotope{N})::Vector{Vector{N}} where {N<:Real}

Return the vertices of a zonotope.

Input

Z – zonotope

Output

List of vertices.

Notes

This implementation computes a convex hull.

For high dimensions, it would be preferable to develop a vertex_iterator approach.

source

LazySets.order — Method.

order(Z::Zonotope)::Rational

Return the order of a zonotope.

Input

Z – zonotope

Output

A rational number representing the order of the zonotope.

Notes

The order of a zonotope is defined as the quotient of its number of generators and its dimension.

source