Singleton

LazySets.Singleton — Type

Singleton{N, VN<:AbstractVector{N}} <: AbstractSingleton{N}

Type that represents a singleton, that is, a set with a unique element.

Fields

element – the only element of the set

Base.rand — Method

rand(::Type{Singleton}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Create a random singleton.

Input

Singleton – type for dispatch

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

A random singleton.

Algorithm

The element is a normally distributed vector with entries of mean 0 and standard deviation 1.

LazySets.element — Method

element(S::Singleton)

Return the element of a singleton.

Input

S – singleton

Output

The element of the singleton.

LazySets.element — Method

element(S::Singleton, i::Int)

Return the i-th entry of the element of a singleton.

Input

S – singleton

i – dimension

Output

The i-th entry of the element of the singleton.

LazySets.translate — Method

translate(S::Singleton, v::AbstractVector)

Translate (i.e., shift) a singleton by a given vector.

Input

S – singleton

v – translation vector

Output

A translated singleton.

Algorithm

We add the vector to the point in the singleton.

LazySets.Arrays.rectify — Method

rectify(S::Singleton)

Concrete rectification of a singleton.

Input

S – singleton

Output

The Singleton that corresponds to the rectification of S.

LazySets.project — Method

project(S::Singleton, block::AbstractVector{Int})

Concrete projection of a singleton.

Input

S – singleton

block – block structure, a vector with the dimensions of interest

Output

A set representing the projection of the singleton S on the dimensions specified by block.

Inherited from LazySet:

diameter

singleton_list

Inherited from AbstractPolytope:

isbounded

isuniversal

Inherited from AbstractCentrallySymmetricPolytope:

dim

isempty

an_element

Inherited from AbstractZonotope:

ngens

order

togrep

Inherited from AbstractHyperrectangle:

norm

radius

high

low

constraints_list

Inherited from AbstractSingleton:

σ

∈

center

vertices

vertices_list

radius_hyperrectangle

radius_hyperrectangle

linear_map

generators

genmat