Tetrahedron

LazySets.TetrahedronModule.Tetrahedron — Type

Tetrahedron{N, VN<:AbstractVector{N}, VT<:AbstractVector{VN}} <: AbstractPolytope{N}

Type that represents a (3-dimensional) tetrahedron in vertex representation.

Fields

vertices – list of vertices

Examples

A tetrahedron can be constructed by passing the list of vertices. The following builds the tetrahedron with edge length 2 from the wikipedia page Tetrahedron:

julia> vertices = [[1, 0, -1/sqrt(2)], [-1, 0, -1/sqrt(2)], [0, 1, 1/sqrt(2)], [0, -1, 1/sqrt(2)]];

julia> T = Tetrahedron(vertices);

julia> dim(T)
3

julia> zeros(3) ∈ T
true

julia> σ(ones(3), T)
3-element Vector{Float64}:
 0.0
 1.0
 0.7071067811865475

Operations

Base.:∈ — Method

∈(x::AbstractVector, X::LazySet)

Check whether a point lies in a set.

Input

x – point/vector

X – set

Output

true iff $x ∈ X$.

Base.:∈ — Method

Extended help

∈(x::AbstractVector, T::Tetrahedron)

Algorithm

For each plane of the tetrahedron, we check if the point x is on the same side as the remaining vertex. We need to check this for each plane.

LazySets.API.σ — Method

σ(d::AbstractVector, X::LazySet)

Compute a support vector of a set in a given direction.

Input

d – direction

X – set

Output

A support vector of X in direction d.

Notes

The convenience alias support_vector is also available.

LazySets.API.σ — Method

Extended help

σ(d::AbstractVector, T::Tetrahedron)

Algorithm

This method falls back to the VPolytope implementation.

Undocumented implementations:

constraints_list

dim

isoperationtype

rand

Inherited from LazySet:

complement

area

chebyshev_center_radius

concretize

constraints

convex_hull

copy(::Type{LazySet})

diameter

eltype

eltype

isoperation

norm

polyhedron

radius

rationalize

rectify

reflect

singleton_list

tosimplehrep

triangulate

triangulate_faces

vertices

affine_map

exponential_map

is_interior_point

linear_map

sample

scale

ρ

translate

cartesian_product

convex_hull

exact_sum

isapprox

==

isequivalent

⊂

minkowski_difference

Inherited from ConvexSet:

isconvextype

linear_combination

Inherited from AbstractPolyhedron:

an_element

extrema

extrema

high

high

low

low

vertices_list

project

intersection

isdisjoint

Inherited from AbstractPolytope:

isbounded

isboundedtype

isempty

ispolyhedral

isuniversal

volume

⊆

minkowski_sum