Utility Functions

Utility functions

LazySets.sign_cadlag — Function.

sign_cadlag(x::N)::N where {N<:Real}

This function works like the sign function but is $1$ for input $0$.

Input

x – real scalar

Output

$1$ if $x ≥ 0$, $-1$ otherwise.

Notes

This is the sign function right-continuous at zero (see càdlàg function). It can be used with vector-valued arguments via the dot operator.

Examples

julia> import LazySets.sign_cadlag

julia> sign_cadlag.([-0.6, 1.3, 0.0])
3-element Array{Float64,1}:
 -1.0
  1.0
  1.0

source

LazySets.@neutral — Macro.

@neutral(SET, NEUT)

Creates functions to make a set type behave commutative with a given neutral element set type.

Input

SET – set type

Output

Nothing.

Notes

This macro generates three functions (and possibly two more if @absorbing has been used in advance).

Examples

@neutral(MinkowskiSum, N) creates the following functions:

MinkowskiSum(X, N) = X

MinkowskiSum(N, X) = X

MinkowskiSum(N, N) = N

source

LazySets.@absorbing — Macro.

@absorbing(SET, ABS)

Creates functions to make a set type behave commutative with a given absorbing element set type.

Input

SET – set type

Output

Nothing.

Notes

This macro generates three functions (and possibly two more if @absorbing has been used in advance).

Examples

@absorbing(MinkowskiSum, A) creates the following functions:

MinkowskiSum(X, A) = A

MinkowskiSum(A, X) = A

MinkowskiSum(A, A) = A

source