Iterative refinement

Iterative Refinement

This section of the manual describes an approximation method for an arbitrary two-dimensional convex set $S$ and a given error bound $ɛ$ using support vectors.

The basic idea is to add new supporting directions whenever the approximation error is still bigger than $ɛ$.

The approximation is represented by a list of local refinements. Each refinement describes a set with one angle and is wrapped in the following type.

LazySets.Approximations.Approximation2D — Type.

Approximation2D{N<:AbstractFloat}

Type that represents a local approximation in 2D.

Fields

p1 – first inner point

d1 – first direction

p2 – second inner point

d2 – second direction

err – error made

ndir – normal direction of the inner approximation

refinable – states if this approximation is refinable

source

LazySets.Approximations.Approximation2D — Method.

Approximation2D(p1::Vector{N}, d1::Vector{N}, p2::Vector{N}, d2::Vector{N}) where {N<:AbstractFloat}

Constructor of Approximation2D from two inner points and two directions.

Input

p1 – first inner point

d1 – first direction

p2 – second inner point

d2 – second direction

Output

A new Approximation2D instance.

source

The approximation is initialized with box directions, i.e., we have four refinement instances, one for each angle. Then we just iterate through all refinement instances and check if the error is bigger than the threshold individually. If so, we refine the instance by splitting into two more precise refinement instances and apply the checks recursively.

LazySets.Approximations.refine — Method.

refine(S::LazySet, approx::Approximation2D)::Tuple{Approximation2D, Approximation2D}

Refine the given approximation.

Input

S – 2D convex set that is approximated

approx – approximation to refine

Output

The refined approximation.

source