Approximations

Approximations

This section of the manual describes the Cartesian decomposition algorithms and the approximation of high-dimensional convex sets using projections.

Approximations

Cartesian Decomposition

Box Approximations

Metric properties of sets

Iterative refinement

LazySets.Approximations — Module.

Module Approximations.jl – polygonal approximation of convex sets through support vectors.

source

Cartesian Decomposition

LazySets.Approximations.decompose — Function.

decompose(S::LazySet)::CartesianProductArray

Compute an overapproximation of the projections of the given convex set over each two-dimensional subspace using box directions.

Input

S – convex set

Output

A CartesianProductArray corresponding to the Cartesian product of $2 \times 2$ box-shaped polygons.

source

decompose(S::LazySet, ɛi::Vector{Float64})::CartesianProductArray

Compute an overapproximation of the projections of the given convex set over each two-dimensional subspace with a certified error bound.

Input

S – convex set

ɛi – array with the error bound for each projection (different error bounds can be passed for different blocks)

Output

A CartesianProductArray corresponding to the Cartesian product of $2 \times 2$ polygons.

Algorithm

This algorithm assumes a decomposition into two-dimensional subspaces only, i.e., partitions of the form $[2, 2, …, 2]$. In particular, if S is a CartesianProductArray, no check is performed to verify that assumption.

The algorithm proceeds as follows:

Project the set S into each partition, with M⋅S, where M is the identity matrix in the block coordinates and zero otherwise.

Overapproximate the set with a given error bound, ɛi[i], for $i = 1, …, b$,

Return the result as a CartesianProductArray.

source

decompose(S::LazySet, ɛ::Float64)::CartesianProductArray

Compute an overapproximation of the projections of the given convex set over each two-dimensional subspace with a certified error bound.

Input

S – convex set

ɛ – error bound

Output

A CartesianProductArray corresponding to the Cartesian product of $2 \times 2$ polygons.

Notes

This function is a particular case of decompose(S, ɛi), where the same error bound for each block is assumed.

source

LazySets.Approximations.overapproximate — Function.

overapproximate(S::LazySet)::HPolygon

Return an approximation of a given 2D convex set as a box-shaped polygon.

Input

S – convex set, assumed to be two-dimensional

Output

A box-shaped polygon in constraint representation.

source

overapproximate(S::LazySet, ɛ::Float64)::HPolygon

Return an ɛ-close approximation of the given 2D set (in terms of Hausdorff distance) as a polygon.

Input

S – convex set, assumed to be two-dimensional

ɛ – error bound

Output

A polygon in constraint representation.

source

Box Approximations

LazySets.Approximations.ballinf_approximation — Function.

ballinf_approximation(S)

Overapproximate a convex set by a tight ball in the infinity norm.

Input

S – convex set

Output

A tight ball in the infinity norm.

Algorithm

The center and radius of the box are obtained by evaluating the support function of the given convex set along the canonical directions.

source

LazySets.Approximations.box_approximation — Function.

box_approximation(S::LazySet)::Hyperrectangle

Overapproximate a convex set by a tight hyperrectangle.

Input

S – convex set

Output

A tight hyperrectangle.

Algorithm

The center of the hyperrectangle is obtained by averaging the support function of the given set in the canonical directions, and the lengths of the sides can be recovered from the distance among support functions in the same directions.

source

LazySets.Approximations.interval_hull — Function.

interval_hull

Alias for box_approximation.

source

LazySets.Approximations.box_approximation_symmetric — Function.

box_approximation_symmetric(S::LazySet)::Hyperrectangle

Overapproximate a convex set by a tight hyperrectangle centered in the origin.

Input

S – convex set

Output

A tight hyperrectangle centered in the origin.

Algorithm

The center of the box is the origin, and the radius is obtained by computing the maximum value of the support function evaluated at the canonical directions.

source

LazySets.Approximations.symmetric_interval_hull — Function.

symmetric_interval_hull

Alias for box_approximation_symmetric.

source

LazySets.Approximations.box_approximation_helper — Function.

box_approximation_helper(S::LazySet)

Common code of box_approximation and box_approximation_symmetric.

Input

S – convex set

Output

A tuple containing the data that is needed to construct a tightly overapproximating hyperrectangle.

c – center

r – radius

Algorithm

The center of the hyperrectangle is obtained by averaging the support function of the given convex set in the canonical directions. The lengths of the sides can be recovered from the distance among support functions in the same directions.

source

Metric properties of sets

Base.LinAlg.norm — Function.

norm(S::LazySet, [p]::Real=Inf)

Return the norm of a convex set. It is the norm of the enclosing ball (of the given norm) of minimal volume.

Input

S – convex set

p – (optional, default: Inf) norm

Output

A real number representing the norm.

source

LazySets.radius — Function.

radius(S::LazySet, [p]::Real=Inf)

Return the radius of a convex set. It is the radius of the enclosing ball (of the given norm) of minimal volume with the same center.

Input

S – convex set

p – (optional, default: Inf) norm

Output

A real number representing the radius.

source

LazySets.diameter — Function.

diameter(S::LazySet, [p]::Real=Inf)

Return the diameter of a convex set. It is the maximum distance between any two elements of the set, or, equivalently, the diameter of the enclosing ball (of the given norm) of minimal volume with the same center.

Input

S – convex set

p – (optional, default: Inf) norm

Output

A real number representing the diameter.

source

Iterative refinement

LazySets.Approximations.approximate — Function.

approximate(S::LazySet, ɛ::Float64)::Vector{Approximation2D}

Return an ɛ-close approximation of the given 2D convex set (in terms of Hausdorff distance) as an inner and an outer approximation composed by sorted local Approximation2D.

Input

S – 2D convex set

ɛ – error bound

Output

An ɛ-close approximation of the given 2D convex set.

source

See Iterative Refinement for more details.