Bloating

Bloating

LazySets.Bloating — Type.

Bloating{N<:AbstractFloat, S<:LazySet{N}} <:LazySet{N}

Type that represents a uniform expansion of a convex set in a given norm (also known as bloating).

Fields

X – convex set

ε – (positive) bloating factor

p – $p$-norm ($≥ 1$; default: $2$)

Notes

Bloating(X, ε, p) is equivalent to the Minkowski sum of X and a ball in the p-norm of radius ε centered in the origin O (i.e., X ⊕ Ballp(p, O, ε)).

source

LazySets.dim — Method.

dim(B::Bloating)

Return the dimension of a bloated convex set.

Input

B – bloated convex set

Output

The ambient dimension of the bloated set.

source

LazySets.σ — Method.

σ(d::AbstractVector{N}, B::Bloating{N}) where {N<:AbstractFloat}

Return the support vector of a bloated convex set in a given direction.

Input

d – direction

B – bloated convex set

Output

The support vector of the bloated set in the given direction.

source

LazySets.ρ — Method.

ρ(d::AbstractVector{N}, B::Bloating{N}) where {N<:AbstractFloat}

Return the support function of a bloated convex set in a given direction.

Input

d – direction

B – bloated convex set

Output

The support function of the bloated set in the given direction.

source

LazySets.isbounded — Method.

isbounded(B::Bloating)

Determine whether a bloated convex set is bounded.

Input

B – bloated convex set

Output

true iff the wrapped set is bounded.

source

Base.isempty — Method.

isempty(B::Bloating)

Determine whether a bloated convex set is empty.

Input

B – bloated convex set

Output

true iff the wrapped set is empty.

source

LazySets.an_element — Method.

an_element(B::Bloating)

Return some element of a bloated convex set.

Input

B – bloated convex set

Output

An element in the bloated convex set.

Algorithm

The implementation returns the result of an_element for the wrapped set.

source