Common Set Operations

Common Set Operations

This section of the manual describes the basic symbolic types describing operations between sets.

Common Set Operations

Minkowski Sum

Convex Hull

Cartesian Product

Linear Maps

Exponential Maps

Minkowski Sum

LazySets.MinkowskiSum — Type.

MinkowskiSum <: LazySet

Type that represents the Minkowski sum of two convex sets.

Fields

X – a convex set

Y – a convex set

source

LazySets.MinkowskiSumArray — Type.

MinkowskiSumArray <: LazySet

Type that represents the Minkowski sum of a finite number of sets.

Fields

sfarray – array of sets

Notes

This type is optimized to be used on the left-hand side of additions only.

source

Convex Hull

LazySets.ConvexHull — Type.

ConvexHull <: LazySet

Type that represents the convex hull of the union of two convex sets.

Fields

X – a convex set

Y – another convex set

source

Cartesian Product

LazySets.CartesianProduct — Type.

CartesianProduct <: LazySet

Type that represents the cartesian product.

Fields

X – convex set

Y – another convex set

For the cartesian product a several sets, there exists a special type CartesianProductArray.

source

LazySets.CartesianProductArray — Type.

CartesianProductArray <: LazySet

Type that represents the cartesian product of a finite number of sets.

Fields

sfarray – array of sets

source

Linear Maps

LazySets.LinearMap — Type.

LinearMap <: LazySet

Type that represents a linear transform of a set. This class is a wrapper around a linear transformation $M⋅S$ of a set $S$, such that it changes the behaviour of the support vector of the new set.

Fields

M – a linear map, which can a be densem matrix, sparse matrix or a subarray object

sf – a convex set represented by its support function

source

Exponential Maps

LazySets.ExponentialMap — Type.

ExponentialMap <: LazySet

Type that represents the action of an exponential map on a set.

Fields

spmexp – a matrix exponential

X – a convex set represented by its support function

source

LazySets.ExponentialProjectionMap — Type.

ExponentialProjectionMap

Type that represents the application of the projection of a SparseMatrixExp over a given set.

Fields

spmexp – the projection of an exponential map

X – a set represented by its support function

source

LazySets.ProjectionSparseMatrixExp — Type.

ProjectionSparseMatrixExp

Type that represents the projection of a SparseMatrixExp.

Fields

L – left multiplication matrix

E – the exponential of a sparse matrix

R – right multiplication matrix

Output

A type that abstract the matrix operation L * exp(E.M) * R, for a given sparse matrix E.M.

source

LazySets.SparseMatrixExp — Type.

SparseMatrixExp

Type that represents the matrix exponential of a sparse matrix, and provides evaluation of its action on vectors.

Fields

M – sparse matrix

Notes

This class is provided for use with very large and very sparse matrices. The evaluation of the exponential matrix action over vectores relies on the Expokit package.

source

LazySets.σ — Method.

σ(d, eprojmap)

Support vector of an ExponentialProjectionMap.

Input

d – a direction

eprojmap – the projection of an exponential map

If S = (LMR)B, where L and R are dense matrices, M is a matrix exponential, and B is a set, it follows that: σ(d, S) = LMR σ(R^T M^T L^T d, B) for any direction d.

source