Common Set Operations

Common Set Operations

This section of the manual describes the basic symbolic types describing operations between sets.

Common Set Operations

Minkowski Sum

Convex Hull

Convex Hull Algorithms

Cartesian Product

Linear Maps

Exponential Maps

Minkowski Sum

LazySets.MinkowskiSum — Type.

MinkowskiSum <: LazySet

Type that represents the Minkowski sum of two convex sets.

Fields

X – a convex set

Y – a convex set

source

LazySets.dim — Method.

dim(ms)

Ambient dimension of a Minkowski product.

Input

ms – Minkowski sum

source

LazySets.σ — Method.

σ(d::AbstractVector{Float64}, ms::MinkowskiSum)

Support vector of a Minkowski sum.

Input

d – vector

ms – Minkowski sum

source

LazySets.MinkowskiSumArray — Type.

MinkowskiSumArray <: LazySet

Type that represents the Minkowski sum of a finite number of sets.

Fields

sfarray – array of sets

Notes

This type is optimized to be used on the left-hand side of additions only.

source

LazySets.dim — Method.

dim(ms::MinkowskiSumArray)

Ambient dimension of the Minkowski sum of a finite number of sets.

Input

ms – Minkowski sum array

Notes

We do not double-check that the dimensions always match.

source

LazySets.σ — Method.

σ(d::Vector{Float64}, ms::MinkowskiSumArray)

Support vector of the Minkowski sum of a finite number of sets.

Input

d – direction

ms – Minkowski sum array

source

Base.:+ — Method.

+(msa, sf)

Adds the support function to the array.

Input

msa – Minkowski sum array

sf – general support function

Notes

This function is overridden for more specific types of sf.

source

Base.:+ — Method.

+(msa1, msa2)

Appends the elements of the second array to the first array.

Input

msa1 – first Minkowski sum array

msa2 – second Minkowski sum array

source

Base.:+ — Method.

+(msa, vs)

Returns the original array because addition with a void set is a no-op.

Input

msa – Minkowski sum array

vs – void set

source

Convex Hull

LazySets.ConvexHull — Type.

ConvexHull <: LazySet

Type that represents the convex hull of the union of two convex sets.

Fields

X – a convex set

Y – another convex set

source

LazySets.dim — Method.

dim(P)

Return the ambient dimension of the convex hull of two sets.

Input

ch – the convex hull of two sets

source

LazySets.σ — Method.

σ(d, P)

Return the support vector of a convex hull in a given direction.

Input

d – direction

ch – the convex hull of two sets

source

Convex Hull Algorithms

LazySets.convex_hull — Function.

convex_hull(points; algorithm)

Compute the convex hull of points in the plane.

Input

points – array of vectors containing the 2D coordinates of the points

algorithm – (optional, default: "monotone_chain") choose the convex hull algorithm, valid options are:

"monotone_chain"

Output

The convex hull as a list of 2D vectors with the coordinates of the points.

Examples

Compute the convex hull of a random set of points:

julia> points = [randn(2) for i in 1:30]; # 30 random points in 2D
julia> hull = convex_hull(points);
julia> typeof(hull)
Array{Array{Float64,1},1}

We can plot the random points, and the polygon whose vertices are the computed convex hull, using Plots:

julia> using Plots
julia> plot([Tuple(pi) for pi in points], seriestype=:scatter)
julia> plot!(VPolygon(hull), alpha=0.2)

source

LazySets.convex_hull! — Function.

convex_hull!(points; algorithm)

Compute the convex hull of points in the plane, in-place. See also: convex_hull.

source

LazySets.right_turn — Function.

right_turn(O, A, B)

Determine if the acute angle defined by the three points O, A, B in the plane is a right turn (counter-clockwise) with respect to the center O.

Input

O – center point

A – one point

B – another point

Algorithm

The cross product is used to determine the sense of rotation. If the result is 0, the points are collinear; if it is positive, the three points constitute a positive angle of rotation around O from A to B; otherwise a negative angle.

source

LazySets.monotone_chain! — Function.

monotone_chain!(points)

Compute the convex hull of points in the plane using Andrew's monotone chain method.

Input

points – array of vectors containing the 2D coordinates of the points; is sorted in-place inside this function

Output

Array of vectors containing the 2D coordinates of the corner points of the convex hull.

Notes

For large sets of points, it is convenient to use static vectors to get maximum performance. For information on how to convert usual vectors into static vectors, see the type SVector provided by the StaticArrays package.

Algorithm

This function implements Andrew's monotone chain convex hull algorithm to construct the convex hull of a set of $n$ points in the plane in $O(n \log n)$ time. For further details see the wikipedia page: Monotone chain

source

Cartesian Product

LazySets.CartesianProduct — Type.

CartesianProduct <: LazySet

Type that represents the cartesian product.

Fields

X – convex set

Y – another convex set

For the cartesian product a several sets, there exists a special type CartesianProductArray.

source

LazySets.dim — Method.

dim(cp)

Ambient dimension of a Cartesian product.

Input

cp – cartesian product

source

LazySets.σ — Method.

σ(d, cp)

Support vector of a Cartesian product.

Input

d – direction

cp – cartesian product

source

LazySets.is_contained — Method.

is_contained(d, cp)

Return whether a vector belongs to a given cartesian product set.

Input

d – a vector

cp – a cartesian product

Output

Return true iff d ∈ cp.

source

LazySets.CartesianProductArray — Type.

CartesianProductArray <: LazySet

Type that represents the cartesian product of a finite number of sets.

Fields

sfarray – array of sets

source

LazySets.dim — Method.

dim(cp)

Ambient dimension of the Cartesian product of a finite number of sets.

Input

cp – cartesian product array

source

LazySets.σ — Method.

σ(d, cp)

Support vector of the Cartesian product of a finite number of sets.

Input

d – direction

cp – cartesian product array

source

LazySets.is_contained — Method.

is_contained(d, cp)

Return whether a given vector is contained in the cartesian product of a finite number of sets.

Input

d – vector

cp – cartesian product array

source

Linear Maps

LazySets.LinearMap — Type.

LinearMap <: LazySet

Type that represents a linear transform of a set. This class is a wrapper around a linear transformation $M⋅S$ of a set $S$, such that it changes the behaviour of the support vector of the new set.

Fields

M – a linear map, which can a be densem matrix, sparse matrix or a subarray object

sf – a convex set represented by its support function

source

LazySets.dim — Method.

dim(lm)

Ambient dimension of the linear map of a set.

It corresponds to the output dimension of the linear map.

Input

lm – a linear map

source

LazySets.σ — Method.

σ(d, lm)

Support vector of the linear map of a set.

If S = MB, where M is sa matrix and B is a set, it follows that σ(d, S) = Mσ(M^T d, B) for any direction d.

Input

d – a direction

lm – a linear map

source

Exponential Maps

LazySets.ExponentialMap — Type.

ExponentialMap <: LazySet

Type that represents the action of an exponential map on a set.

Fields

spmexp – a matrix exponential

X – a convex set represented by its support function

source

LazySets.dim — Method.

dim(em)

The ambient dimension of a ExponentialMap.

Input

em – an ExponentialMap

source

LazySets.σ — Method.

σ(d, eprojmap)

Support vector of an ExponentialProjectionMap.

Input

d – a direction

eprojmap – the projection of an exponential map

If S = (LMR)B, where L and R are dense matrices, M is a matrix exponential, and B is a set, it follows that: σ(d, S) = LMR σ(R^T M^T L^T d, B) for any direction d.

source

LazySets.ExponentialProjectionMap — Type.

ExponentialProjectionMap

Type that represents the application of the projection of a SparseMatrixExp over a given set.

Fields

spmexp – the projection of an exponential map

X – a set represented by its support function

source

LazySets.dim — Method.

dim(eprojmap)

The ambient dimension of a ExponentialProjectionMap.

It is given by the output dimension (left-most matrix).

Input

eprojmap – an ExponentialProjectionMap

source

LazySets.ProjectionSparseMatrixExp — Type.

ProjectionSparseMatrixExp

Type that represents the projection of a SparseMatrixExp.

Fields

L – left multiplication matrix

E – the exponential of a sparse matrix

R – right multiplication matrix

Output

A type that abstract the matrix operation L * exp(E.M) * R, for a given sparse matrix E.M.

source

LazySets.SparseMatrixExp — Type.

SparseMatrixExp

Type that represents the matrix exponential of a sparse matrix, and provides evaluation of its action on vectors.

Fields

M – sparse matrix

Notes

This class is provided for use with very large and very sparse matrices. The evaluation of the exponential matrix action over vectores relies on the Expokit package.

source