Common Set Representations

Common Set Representations

This section of the manual describes the basic set representation types.

Common Set Representations

Balls

Polygons

Lines and linear constraints

VoidSet

Singleton

Balls

Unit balls are defined by int center (vector) and radius (scalar), such as infinity-norm balls,

\[B_\infty(c, r) = \{ x ∈ \mathbb{R}^n : \Vert x - c\Vert_\infty \leq r \}.\]

and Euclidean (2-norm) balls,

\[B_2(c, r) = \{ x ∈ \mathbb{R}^n : \Vert x - c\Vert_2 \leq r \}.\]

LazySets.Hyperrectangle — Type.

Hyperrectangle <: LazySet

Type that represents a Hyperrectangle.

A hyperrectangle is the Cartesian product of one-dimensional intervals.

Fields

center – center of the hyperrectangle as a real vector

radius – radius of the ball as a real vector, i.e. its width along each coordinate direction

source

LazySets.BallInf — Type.

BallInf <: LazySet

Type that represents a ball in the infinity norm.

Fields

center – center of the ball as a real vector

radius – radius of the ball as a scalar ($≧ 0$)

Examples

We create the two-dimensional unit ball, and compute its support function along the direction $(1, 1)$:

julia> B = BallInf(zeros(2), 0.1)
LazySets.BallInf([0.0, 0.0], 0.1)

julia> dim(B)
2

julia> ρ([1., 1.], B)
0.2

source

LazySets.Ball2 — Type.

Ball2 <: LazySet

Type that represents a ball in the 2-norm.

Fields

center – center of the ball as a real vector

radius – radius of the ball as a scalar ($≧ 0$)

Examples

A five-dimensional ball in the 2-norm centered at the origin of radius 0.5:

julia> using LazySets
julia> B = Ball2(zeros(5), 0.5)
LazySets.Ball2([0.0, 0.0, 0.0, 0.0, 0.0], 0.5)
julia> dim(B)
5

We evaluate the support vector in a given direction:

julia> σ(ones(5), B)
5-element Array{Float64,1}:
0.06742
0.13484
0.20226
0.26968
0.3371

source

Polygons

LazySets.HPolygon — Type.

HPolygon <: LazySet

Type that represents a convex polygon (in H-representation).

Fields

constraints – an array of linear constraints

source

LazySets.VPolygon — Type.

VPolygon

Type that represents a polygon by its vertices.

Fields

vl – the list of vertices

source

LazySets.plot_polygon — Function.

plot_polygon(P, backend, [name], [gridlines])

Plot a polygon given in constraint form.

Input

P – a polygon, given as a HPolygon or the refined class HPolygonOpt

backend – (optional, default: $'pyplot'$): select the plot backend; valid options are:

pyplot_savefig – use PyPlot package, save to a file

pyplot_inline – use PyPlot package, showing in external program

gadfly – use Gadfly package, showing in browser

'' – (empty string), return nothing, without plotting

name – (optional, default: plot.png) the filename of the plot, if it is saved to disk

gridlines – (optional, default: false) to display or not gridlines in the output plot

Examples

This function can receive one polygon, as in:

julia> using LazySets, PyPlot
julia> H = HPolygon([LinearConstraint([1.0, 0.0], 0.6), LinearConstraint([0.0, 1.0], 0.6),
       LinearConstraint([-1.0, 0.0], -0.4), LinearConstraint([0.0, -1.0], -0.4)])
julia> plot_polygon(H, backend="pyplot_inline");

Multiple polygons can be plotted passing a list instead of a single element:

julia> Haux = HPolygon([LinearConstraint([1.0, 0.0], 1.2), LinearConstraint([0.0, 1.0], 1.2),
       LinearConstraint([-1.0, 0.0], -0.8), LinearConstraint([0.0, -1.0], -0.8)])
julia> plot_polygon([H, Haux], backend="pyplot_inline");

source

LazySets.tovrep — Function.

tovrep(s)

Build a vertex representation of the given polygon.

Input

s – a polygon in H-representation, HPolygon. The linear constraints are assumed sorted by their normal directions.

Output

The same polygon in a vertex representation, VPolygon.

source

tovrep(po)

Build a vertex representation of the given polygon.

Input

po – a polygon in H-representation. The linear constraints are assumed sorted by their normal directions.

Output

The same polygon in a vertex representation.

source

Lines and linear constraints

LazySets.intersection — Function.

intersection(Δ1, Δ2)

Return the intersection of two 2D lines.

Input

Δ1 – a line

Δ2 – another line

Output

The intersection point.

Examples

The line y = -x + 1 intersected with y = x:

julia> intersection(Line([1., 1.], 1.), Line([-1., 1.], 0.))
2-element Array{Float64,1}:
 0.5
 0.5

source

LazySets.LinearConstraint — Type.

LinearConstraint

Type that represents a linear constraint (a half-space) of the form a⋅x ≦ b.

Fields

a – a normal direction

b – the constraint

EXAMPLES:

The set y >= 0 in the plane:

julia> LinearConstraint([0, -1.], 0.)
LazySets.LinearConstraint([0.0, -1.0], 0.0)

source

LazySets.Line — Type.

Line

Type that represents a line in 2D of the form a⋅x = b.

Fields

a – a normal direction (size = 2)

b – the constraint

Examples

The line y = -x + 1:

julia> Line([1., 1.], 1.)
LazySets.Line([1.0, 1.0], 1.0)

source

VoidSet

LazySets.VoidSet — Type.

VoidSet <: LazySet

Type that represents a void (neutral) set with respect to Minkowski sum.

Fields

dim – ambient dimension of the VoidSet

source

Singleton

LazySets.Singleton — Type.

Singleton <: LazySet

Type that represents a singleton, that is, a set with a unique element.

Fields

element – the only element of the set

source