Common Set Representations

Common Set Representations

This section of the manual describes the basic set representation types.

Common Set Representations

Abstract support function and support vector

Balls

Polygons

Lines and linear constraints

Hyperrectangles

VoidSets

Singletons

Abstract support function and support vector

LazySets — Module.

Main module for LazySets.jl – a Julia package for calculus with convex sets.

source

LazySets.LazySet — Type.

LazySet

Abstract type for a lazy set.

Every concrete LazySet must define a σ(d, X), representing the support vector of X in a given direction d, and dim, the ambient dimension of the set X.

source

LazySets.ρ — Function.

ρ(d::Vector{Float64}, sf::LazySet)::Float64

Evaluate the support function of a set in a given direction.

Input

d – a real vector, the direction investigated

sf – a convex set

Output

ρ(d, sf) – the support function

source

LazySets.σ — Function.

σ(d, B)

Return the support vector of a Ball2 in a given direction.

Input

d – a direction

B – a ball in the 2-norm

Output

The support vector in the given direction.

Notes

If the given direction has norm zero, the origin is returned.

source

σ(d, B)

Return the support vector of an infinity-norm ball in a given direction.

Input

d – direction

B – unit ball in the infinity norm

Algorithm

This code is a vectorized version of

[(d[i] >= 0) ? B.center[i] + B.radius : B.center[i] - B.radius for i in 1:length(d)]

Notice that we cannot use B.center + sign.(d) * B.radius, since the built-in sign function is such that sign(0) = 0, instead of 1. For this reason, we use the custom unit_step function, that allows to do: B.center + unit_step.(d) * B.radius (the dot operator performs broadcasting, to accept vector-valued entries).

source

σ(d, H)

Return the support vector of a Hyperrectangle in a given direction.

source

σ(d, P)

Return the support vector of a polygon in a given direction. Return the zero vector if there are no constraints (i.e., the universal polytope).

Input

d – direction

p – polyhedron in H-representation

source

σ(d, P)

Return the support vector of the optimized polygon in a given direction.

Input

d – direction

P – polyhedron in H-representation

source

σ(d, P)

Return the support vector of a convex hull in a given direction.

Input

d – direction

ch – the convex hull of two sets

source

σ(d, cp)

Support vector of a Cartesian product.

Input

d – direction

cp – cartesian product

source

σ(d, cp)

Support vector of the Cartesian product of a finite number of sets.

Input

d – direction

cp – cartesian product array

source

σ(d, eprojmap)

Support vector of an ExponentialProjectionMap.

Input

d – a direction

eprojmap – the projection of an exponential map

If S = (LMR)B, where L and R are dense matrices, M is a matrix exponential, and B is a set, it follows that: σ(d, S) = LMR σ(R^T M^T L^T d, B) for any direction d.

source

σ(d, lm)

Support vector of the linear map of a set.

If S = MB, where M is sa matrix and B is a set, it follows that σ(d, S) = Mσ(M^T d, B) for any direction d.

Input

d – a direction

lm – a linear map

source

σ(d::AbstractVector{Float64}, ms::MinkowskiSum)

Support vector of a Minkowski sum.

Input

d – vector

ms – Minkowski sum

source

σ(d::Vector{Float64}, ms::MinkowskiSumArray)

Support vector of the Minkowski sum of a finite number of sets.

Input

d – direction

ms – Minkowski sum array

source

Balls

Unit balls are defined by int center (vector) and radius (scalar), such as infinity-norm balls,

\[B_\infty(c, r) = \{ x ∈ \mathbb{R}^n : \Vert x - c\Vert_\infty \leq r \}.\]

and Euclidean (2-norm) balls,

\[B_2(c, r) = \{ x ∈ \mathbb{R}^n : \Vert x - c\Vert_2 \leq r \}.\]

LazySets.BallInf — Type.

BallInf <: LazySet

Type that represents a ball in the infinity norm.

Fields

center – center of the ball as a real vector

radius – radius of the ball as a scalar ($≧ 0$)

Examples

We create the two-dimensional unit ball, and compute its support function along the direction $(1, 1)$:

julia> B = BallInf(zeros(2), 0.1)
LazySets.BallInf([0.0, 0.0], 0.1)

julia> dim(B)
2

julia> ρ([1., 1.], B)
0.2

source

LazySets.dim — Method.

dim(B)

Return the dimension of a BallInf.

Input

B – a ball in the infinity norm

Output

The ambient dimension of the ball.

source

LazySets.σ — Method.

σ(d, B)

Return the support vector of an infinity-norm ball in a given direction.

Input

d – direction

B – unit ball in the infinity norm

Algorithm

This code is a vectorized version of

[(d[i] >= 0) ? B.center[i] + B.radius : B.center[i] - B.radius for i in 1:length(d)]

Notice that we cannot use B.center + sign.(d) * B.radius, since the built-in sign function is such that sign(0) = 0, instead of 1. For this reason, we use the custom unit_step function, that allows to do: B.center + unit_step.(d) * B.radius (the dot operator performs broadcasting, to accept vector-valued entries).

source

LazySets.Ball2 — Type.

Ball2 <: LazySet

Type that represents a ball in the 2-norm.

Fields

center – center of the ball as a real vector

radius – radius of the ball as a scalar ($≧ 0$)

Examples

A five-dimensional ball in the 2-norm centered at the origin of radius 0.5:

julia> using LazySets
julia> B = Ball2(zeros(5), 0.5)
LazySets.Ball2([0.0, 0.0, 0.0, 0.0, 0.0], 0.5)
julia> dim(B)
5

We evaluate the support vector in a given direction:

julia> σ(ones(5), B)
5-element Array{Float64,1}:
0.06742
0.13484
0.20226
0.26968
0.3371

source

LazySets.dim — Method.

dim(B)

Return the dimension of a Ball2.

Input

B – a ball in the 2-norm

Output

The ambient dimension of the ball.

source

LazySets.σ — Method.

σ(d, B)

Return the support vector of a Ball2 in a given direction.

Input

d – a direction

B – a ball in the 2-norm

Output

The support vector in the given direction.

Notes

If the given direction has norm zero, the origin is returned.

source

Polygons

Constraint representation

LazySets.HPolygon — Type.

HPolygon <: LazySet

Type that represents a convex polygon (in H-representation).

Fields

constraints – an array of linear constraints

source

LazySets.tovrep — Method.

tovrep(s)

Build a vertex representation of the given polygon.

Input

s – a polygon in H-representation, HPolygon. The linear constraints are assumed sorted by their normal directions.

Output

The same polygon in a vertex representation, VPolygon.

source

LazySets.addconstraint! — Method.

addconstraint!(p, c)

Add a linear constraint to a polygon keeping the constraints sorted by their normal directions.

Input

p – a polygon

c – the linear constraint to add

source

LazySets.σ — Method.

σ(d, P)

Return the support vector of a polygon in a given direction. Return the zero vector if there are no constraints (i.e., the universal polytope).

Input

d – direction

p – polyhedron in H-representation

source

LazySets.vertices_list — Method.

vertices_list(po)

Return the list of vertices of a convex polygon.

Input

po – a polygon, which can be either of type HPolygon or the refined type HPolygonOpt

Output

List of vertices as an array of vertex pairs, Vector{Vector{Float64}}.

source

LazySets.is_contained — Method.

is_contained(x, P)

Return whether a given vector is contained in the polygon.

Input

x – vector

P – polygon

Output

Return rue iff x ∈ P.

source

Optimized constraint representation

LazySets.HPolygonOpt — Type.

HPolygonOpt <: LazySet

Type that represents a convex polygon (in H-representation).

Fields

P – polygon

ind – an index in the list of constraints to begin the search to evaluate the support functions.

Notes

This structure is optimized to evaluate the support function/vector with a large sequence of directions, which are one to one close.

source

LazySets.σ — Method.

σ(d, P)

Return the support vector of the optimized polygon in a given direction.

Input

d – direction

P – polyhedron in H-representation

source

LazySets.dim — Method.

dim(P)

Return the ambient dimension of the optimized polygon.

Input

P – optimized polyhedron in H-representation

source

LazySets.tovrep — Method.

tovrep(po)

Build a vertex representation of the given polygon.

Input

po – a polygon in H-representation. The linear constraints are assumed sorted by their normal directions.

Output

The same polygon in a vertex representation.

source

LazySets.is_contained — Method.

is_contained(x, P)

Return whether a given vector is contained in the optimized polygon.

Input

x – vector

P – polygon

Output

Return true iff x ∈ P.

source

Vertex representation

LazySets.VPolygon — Type.

VPolygon <: LazySet

Type that represents a polygon by its vertices.

Fields

vl – the list of vertices

source

Lines and linear constraints

LazySets.intersection — Function.

intersection(Δ1, Δ2)

Return the intersection of two 2D lines.

Input

Δ1 – a line

Δ2 – another line

Output

The intersection point.

Examples

The line y = -x + 1 intersected with y = x:

julia> intersection(Line([1., 1.], 1.), Line([-1., 1.], 0.))
2-element Array{Float64,1}:
 0.5
 0.5

source

LazySets.LinearConstraint — Type.

LinearConstraint

Type that represents a linear constraint (a half-space) of the form a⋅x ≦ b.

Fields

a – a normal direction

b – the constraint

EXAMPLES:

The set y >= 0 in the plane:

julia> LinearConstraint([0, -1.], 0.)
LazySets.LinearConstraint([0.0, -1.0], 0.0)

source

LazySets.Line — Type.

Line

Type that represents a line in 2D of the form a⋅x = b.

Fields

a – a normal direction (size = 2)

b – the constraint

Examples

The line y = -x + 1:

julia> Line([1., 1.], 1.)
LazySets.Line([1.0, 1.0], 1.0)

source

Hyperrectangles

LazySets.Hyperrectangle — Type.

Hyperrectangle <: LazySet

Type that represents a Hyperrectangle.

A hyperrectangle is the Cartesian product of one-dimensional intervals.

Fields

center – center of the hyperrectangle as a real vector

radius – radius of the ball as a real vector, i.e. its width along each coordinate direction

source

LazySets.radius — Method.

radius(H::Hyperrectangle)

Return the radius of a Hyperrectangle. It is the radius of the enclosing hypercube of minimal volume.

Input

H – a hyperrectangle

Output

A real number representing its radius.

source

LazySets.dim — Method.

dim(H)

Return the dimension of a Hyperrectangle.

Input

H – a hyperrectangle

Output

The ambient dimension of the hyperrectangle as an integer.

source

LazySets.σ — Method.

σ(d, H)

Return the support vector of a Hyperrectangle in a given direction.

source

LazySets.vertices_list — Method.

vertices_list(H::Hyperrectangle)

Return the vertices of a hyperrectangle.

Input

H – a hyperrectangle

Output

The list of vertices as an array of floating-point vectors.

Notes

For high-dimensions, it is preferable to develop a vertex_iterator approach.

source

LazySets.diameter — Method.

diameter(H::Hyperrectangle)

Return the diameter of a hyperrectangle. It the maximum norm (measured the infinity norm) of any element of the set.

Input

H – a hyperrectangle

Output

The diameter of the hyperrectangle.

source

VoidSets

LazySets.VoidSet — Type.

VoidSet <: LazySet

Type that represents a void (neutral) set with respect to Minkowski sum.

Fields

dim – ambient dimension of the VoidSet

source

Singletons

LazySets.Singleton — Type.

Singleton <: LazySet

Type that represents a singleton, that is, a set with a unique element.

Fields

element – the only element of the set

source