EmptySet

Empty set (EmptySet)

LazySets.EmptySet — Type.

EmptySet{N<:Real} <: LazySet{N}

Type that represents the empty set, i.e., the set with no elements.

source

LazySets.∅ — Type.

∅

Alias for EmptySet{Float64}.

source

LazySets.dim — Method.

dim(∅::EmptySet)

Return the dimension of an empty set.

Input

∅ – an empty set

Output

The dimension of the empty set.

source

LazySets.σ — Method.

σ(d::AbstractVector{N}, ∅::EmptySet{N}) where {N<:Real}

Return the support vector of an empty set.

Input

d – direction

∅ – empty set

Output

An error.

source

LazySets.ρ — Method.

ρ(d::AbstractVector{N}, ∅::EmptySet{N}) where {N<:Real}

Evaluate the support function of an empty set in a given direction.

Input

d – direction

∅ – empty set

Output

An error.

source

Base.:∈ — Method.

∈(x::AbstractVector{N}, ∅::EmptySet{N}) where {N<:Real}

Check whether a given point is contained in an empty set.

Input

x – point/vector

∅ – empty set

Output

The output is always false.

Examples

julia> [1.0, 0.0] ∈ ∅(2)
false

source

LazySets.an_element — Method.

an_element(∅::EmptySet)

Return some element of an empty set.

Input

∅ – empty set

Output

An error.

source

Base.rand — Method.

rand(::Type{EmptySet}; [N]::Type{<:Real}=Float64, [dim]::Int=0,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Create an empty set (note that there is nothing to randomize).

Input

EmptySet – type for dispatch

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

The (only) empty set of the given numeric type and dimension.

source

LazySets.isbounded — Method.

isbounded(∅::EmptySet)

Determine whether an empty set is bounded.

Input

∅ – empty set

Output

true.

source

Base.isempty — Method.

isempty(∅::EmptySet)

Return if the empty set is empty or not.

Input

∅ – empty set

Output

true.

source

LazySets.isuniversal — Method.

isuniversal(∅::EmptySet{N}, [witness]::Bool=false) where {N<:Real}

Check whether an empty is universal.

Input

∅ – empty set

witness – (optional, default: false) compute a witness if activated

Output

If witness option is deactivated: false

If witness option is activated: (false, v) where $v ∉ S$, although we currently throw an error

source

LinearAlgebra.norm — Function.

norm(S::EmptySet, [p]::Real=Inf)

Return the norm of an empty set. It is the norm of the enclosing ball (of the given $p$-norm) of minimal volume that is centered in the origin.

Input

S – empty set

p – (optional, default: Inf) norm

Output

An error.

source

LazySets.radius — Function.

radius(S::EmptySet, [p]::Real=Inf)

Return the radius of an empty set. It is the radius of the enclosing ball (of the given $p$-norm) of minimal volume with the same center.

Input

S – empty set

p – (optional, default: Inf) norm

Output

An error.

source

LazySets.diameter — Function.

diameter(S::EmptySet, [p]::Real=Inf)

Return the diameter of an empty set. It is the maximum distance between any two elements of the set, or, equivalently, the diameter of the enclosing ball (of the given $p$-norm) of minimal volume with the same center.

Input

S – empty set

p – (optional, default: Inf) norm

Output

An error.

source

LazySets.linear_map — Method.

linear_map(M::AbstractMatrix{N}, ∅::EmptySet{N}) where {N}

Return the linear map of an empty set.

Input

M – matrix

∅ – empty set

Output

The empty set.

source

LazySets.translate — Method.

translate(∅::EmptySet{N}, v::AbstractVector{N}) where {N<:Real}

Translate (i.e., shift) an empty set by a given vector.

Input

∅ – empty set

v – translation vector

Output

The empty set.

source

LazySets.plot_recipe — Method.

plot_recipe(∅::EmptySet{N}, [ε]::N=zero(N)) where {N<:Real}

Convert an empty set to a sequence of points for plotting. In the special case of an empty set, we define the sequence as nothing.

Input

∅ – empty set

ε – (optional, default: 0) ignored, used for dispatch

Output

nothing.

source

RecipesBase.apply_recipe — Method.

plot_emptyset(∅::EmptySet, [ε]::N=zero(N); ...)

Plot an empty set.

Input

∅ – empty set

ε – (optional, default: 0) ignored, used for dispatch

source

Inherited from LazySet:

norm

radius

diameter