Empty set (EmptySet)

LazySets.EmptySet — Type

EmptySet{N} <: ConvexSet{N}

Type that represents the empty set, i.e., the set with no elements.

LazySets.∅ — Type

∅

Alias for EmptySet{Float64}.

LazySets.API.dim — Method

dim(∅::EmptySet)

Return the dimension of an empty set.

Input

∅ – an empty set

Output

The dimension of the empty set.

LazySets.API.σ — Method

σ(d::AbstractVector, ∅::EmptySet)

Return the support vector of an empty set.

Input

d – direction

∅ – empty set

Output

An error.

LazySets.API.ρ — Method

ρ(d::AbstractVector, ∅::EmptySet)

Evaluate the support function of an empty set in a given direction.

Input

d – direction

∅ – empty set

Output

An error.

Base.:∈ — Method

∈(x::AbstractVector, ∅::EmptySet)

Check whether a given point is contained in an empty set.

Input

x – point/vector

∅ – empty set

Output

false.

Examples

julia> [1.0, 0.0] ∈ ∅(2)
false

LazySets.API.an_element — Method

an_element(∅::EmptySet)

Return some element of an empty set.

Input

∅ – empty set

Output

An error.

Base.rand — Method

rand(::Type{EmptySet}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Create an empty set (note that there is nothing to randomize).

Input

EmptySet – type for dispatch

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

The (only) empty set of the given numeric type and dimension.

LazySets.API.isbounded — Method

isbounded(∅::EmptySet)

Check whether an empty set is bounded.

Input

∅ – empty set

Output

true.

Base.isempty — Method

isempty(∅::EmptySet)

Check if the empty set is empty.

Input

∅ – empty set

Output

true.

LazySets.API.isuniversal — Method

isuniversal(∅::EmptySet{N}, [witness]::Bool=false) where {N}

Check whether an empty set is universal.

Input

∅ – empty set

witness – (optional, default: false) compute a witness if activated

Output

If witness option is deactivated: false

If witness option is activated: (false, v) where $v ∉ S$

LazySets.API.vertices — Method

vertices(∅::EmptySet)

Construct an iterator over the vertices of an empty set.

Input

∅ – empty set

Output

The empty iterator, as the empty set does not contain any vertices.

LazySets.API.vertices_list — Method

vertices_list(∅::EmptySet)

Return the list of vertices of an empty set.

Input

∅ – empty set

Output

The empty list of vertices, as the empty set does not contain any vertices.

LinearAlgebra.norm — Function

norm(∅::EmptySet, [p]::Real=Inf)

Return the norm of an empty set. It is the norm of the enclosing ball (of the given $p$-norm) of minimal volume that is centered in the origin.

Input

∅ – empty set

p – (optional, default: Inf) norm

Output

An error.

LazySets.API.radius — Function

radius(∅::EmptySet, [p]::Real=Inf)

Return the radius of an empty set. It is the radius of the enclosing ball (of the given $p$-norm) of minimal volume with the same center.

Input

∅ – empty set

p – (optional, default: Inf) norm

Output

An error.

LazySets.API.diameter — Function

diameter(∅::EmptySet, [p]::Real=Inf)

Return the diameter of an empty set. It is the maximum distance between any two elements of the set or, equivalently, the diameter of the enclosing ball (of the given $p$-norm) of minimal volume with the same center.

Input

∅ – empty set

p – (optional, default: Inf) norm

Output

An error.

LazySets.API.linear_map — Method

linear_map(M::AbstractMatrix{N}, ∅::EmptySet{N}) where {N}

Return the linear map of an empty set.

Input

M – matrix

∅ – empty set

Output

An empty set.

LazySets.API.translate — Method

translate(∅::EmptySet, v::AbstractVector)

Translate (i.e., shift) an empty set by a given vector.

Input

∅ – empty set

v – translation vector

Output

The empty set.

LazySets.plot_recipe — Method

plot_recipe(∅::EmptySet{N}, [ε]=zero(N)) where {N}

Convert an empty set to a sequence of points for plotting. In the special case of an empty set, the sequence is empty.

Input

∅ – empty set

ε – (optional, default: 0) ignored, used for dispatch

Output

An empty array.

RecipesBase.apply_recipe — Method

plot_emptyset(∅::EmptySet, [ε]::Real=zero(N); ...)

Plot an empty set.

Input

∅ – empty set

ε – (optional, default: 0) ignored, used for dispatch

LazySets.API.area — Method

area(∅::EmptySet)

Return the area of an empty set.

Input

∅ – empty set

Output

$0$.

LazySets.API.volume — Method

volume(∅::EmptySet{N}) where {N}

Return the volume of an empty set.

Input

∅ – empty set

Output

$0$.

LazySets.chebyshev_center_radius — Method

chebyshev_center_radius(∅::EmptySet; [kwargs]...)

Compute the Chebyshev center and the corresponding radius of an empty set.

Input

∅ – empty set

kwargs – further keyword arguments (ignored)

Output

An error.

LazySets.API.low — Method

low(∅::EmptySet)

Return a vector with the lowest coordinates of an empty set in each canonical direction.

Input

∅ – empty set

Output

An error.

Notes

See also low(∅::EmptySet, i::Int).

LazySets.API.high — Method

high(∅::EmptySet)

Return a vector with the highest coordinates of an empty set in each canonical direction.

Input

∅ – empty set

Output

An error.

Notes

See also high(∅::EmptySet, i::Int).

LazySets.API.low — Method

low(∅::EmptySet, i::Int)

Return the lowest coordinate of an empty set in the given direction.

Input

∅ – empty set

i – dimension of interest

Output

An error.

LazySets.API.high — Method

high(∅::EmptySet, i::Int)

Return the highest coordinate of an empty set in the given direction.

Input

∅ – empty set

i – dimension of interest

Output

An error.

LazySets.API.complement — Method

complement(∅::EmptySet{N}) where {N}

Return the complement of an empty set.

Input

∅ – empty set

Output

The universe of the same dimension.

complement(U::Universe{N}) where {N}

Return the complement of an universe.

Input

∅ – universe

Output

The empty set of the same dimension.

ReachabilityBase.Arrays.rectify — Method

rectify(∅::EmptySet)

Concrete rectification of an empty set.

Input

∅ – empty set

Output

The empty set.

LazySets.API.reflect — Method

reflect(∅::EmptySet)

Concrete reflection of an empty set.

Input

∅ – empty set

Output

The same empty set.

Inherited from LazySet:

singleton_list