Interval

LazySets.IntervalModule.Interval — Type

Interval{N} <: AbstractHyperrectangle{N}

Type representing an interval on the real line. Mathematically, it is of the form

\[[a, b] := \{ a ≤ x ≤ b \} ⊆ ℝ.\]

Fields

dat – data container for the given interval

Notes

This type relies on the IntervalArithmetic.jl library for representation of intervals and arithmetic operations.

Examples

Unidimensional intervals are symbolic representations of a real closed interval.

We can create intervals in different ways. The simplest way is to pass a pair of numbers:

julia> x = Interval(0.0, 1.0)
Interval{Float64}([0, 1])

A 2-vector is also possible:

julia> x = Interval([0.0, 1.0])
Interval{Float64}([0, 1])

An interval can also be constructed from an IntervalArithmetic.Interval. Note that if the package IntervalArithmetic is loaded in the current scope, you have to prepend the package names to Intervals since there is a name conflict otherwise.

julia> using IntervalArithmetic
WARNING: using IntervalArithmetic.Interval in module Main conflicts with an existing identifier.

julia> x = LazySets.Interval(IntervalArithmetic.Interval(0.0, 1.0))
Interval{Float64}([0, 1])

julia> dim(x)
1

julia> center(x)
1-element Vector{Float64}:
 0.5

The usual pairwise arithmetic operators - and * are interpreted in the standard way known in interval arithmetic, so the results are Intervals. Note that + is generally used for the lazy Minkowksi sum in this library.

Intervals of other numeric types can be created as well, e.g., a rational interval:

julia> Interval(0//1, 2//1)
Interval{Rational{Int64}}([0//1, 2//1])

Conversion

convert(::Type{Interval}, ::LazySet)
convert(::Type{IA.Interval}, ::LazySet)
convert(::Type{Interval}, ::IA.Interval)

Operations

LazySets.API.an_element — Method

an_element(X::LazySet)

Return some element of a nonempty set.

Input

X – set

Output

An element of X unless X is empty.

LazySets.API.an_element — Method

Extended help

an_element(X::Interval)

Algorithm

Return the left border (low(X)) of the interval.

LazySets.chebyshev_center_radius — Method

chebyshev_center_radius(X::Interval; [kwargs]...)

Compute a Chebyshev center and the corresponding radius of an interval.

Input

X – interval

kwargs – further keyword arguments (ignored)

Output

The pair (c, r) where c is the Chebyshev center of X and r is the radius of the largest Euclidean ball with center c enclosed by X.

Notes

The Chebyshev center of an interval is just the center of the interval.

LazySets.isflat — Method

isflat(X::Interval)

Check whether an interval is flat, i.e., whether its extreme values coincide.

Input

X – interval

Output

true if the interval is flat and false otherwise.

Notes

For robustness with respect to floating-point inputs, this function relies on the result of isapproxzero when applied to the diameter of the interval. In other words, this function depends on the absolute zero tolerance ABSZTOL.

LazySets.ngens — Method

ngens(X::Interval)

Return the number of generators of an interval.

Input

X – interval

Output

The number of generators.

Algorithm

An interval has either one generator, or zero generators if it is a degenerated interval of diameter zero.

LazySets.radius_hyperrectangle — Method

radius_hyperrectangle(X::Interval)

Return the box radius of an interval in every dimension.

Input

X – interval

Output

The box radius of the interval (a one-dimensional vector).

LazySets.radius_hyperrectangle — Method

radius_hyperrectangle(X::Interval, i::Int)

Return the box radius of an interval in a given dimension.

Input

X – interval

i – dimension index (must be 1)

Output

The box radius in the given dimension.

Base.rand — Method

rand(T::Type{<:LazySet}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing
    )

Create a random set of the given set type.

Input

T – set type

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

A random set of the given set type.

Base.rand — Method

Extended help

rand(::Type{Interval}; [N]::Type{<:Real}=Float64, [dim]::Int=1,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Notes

All numbers are normally distributed with mean 0 and standard deviation 1.

ReachabilityBase.Arrays.rectify — Method

rectify(X::LazySet)

Compute the rectification of a set.

Input

X – set

Output

A set representing the rectification of X.

ReachabilityBase.Arrays.rectify — Method

Extended help

rectify(X::Interval)

Output

An Interval, even if it represents a singleton only containing the origin (which is the case if the original interval was nonpositive).

LazySets.API.vertices_list — Method

vertices_list(X::LazySet)

Compute a list of vertices of a polytopic set.

Input

X – polytopic set

Output

A list of the vertices of X.

LazySets.API.vertices_list — Method

Extended help

vertices_list(X::Interval; kwargs...)

Output

The list of vertices of the interval, which are two one-dimensional vectors, or just one if the interval is degenerate (the endpoints match within the working tolerance).

LazySets.API.linear_map — Method

linear_map(M::AbstractMatrix, X::LazySet)

Compute the linear map $M · X$.

Input

M – matrix

X – set

Output

A set representing the linear map $M · X$.

LazySets.API.linear_map — Method

Extended help

linear_map(M::AbstractMatrix, X::Interval)

Output

Either an interval or a zonotope, depending on the leading dimension (i.e., the number of rows) of M:

If size(M, 1) == 1, the output is an Interval obtained by scaling X by the matrix M.

If size(M, 1) ≠ 1, the output is a Zonotope with center M * center(X) and the single generator M * g, where g = (high(X)-low(X))/2.

Base.split — Method

split(X::Interval, k)

Partition an interval into k uniform sub-intervals.

Input

X – interval

k – number of sub-intervals, possibly wrapped in a vector of length 1

Output

A list of k Intervals.

LazySets.API.translate — Method

translate(X::LazySet, v::AbstractVector)

Compute the translation of a set with a vector.

Input

X – set

v – vector

Output

A set representing $X + \{v\}$.

LazySets.API.translate — Method

Extended help

translate(X::Interval, v::AbstractVector)

Notes

An in-place version is not available because the IntervalArithmetic.Interval type is immutable.

LazySets.API.difference — Method

difference(X::LazySet, Y::LazySet)

Compute the set difference of two sets.

Input

X – set

Y – set

Output

A set representing the difference $X ∖ Y$.

Notes

The set difference is defined as:

\[ X ∖ Y = \{x \mid x ∈ X \text{ and } x ∉ Y \}\]

LazySets.API.difference — Method

Extended help

difference(X::Interval, Y::Interval)

Output

Depending on the position of the intervals, the output is one of the following:

An EmptySet.

An Interval.

A UnionSet of two Interval sets.

Algorithm

Let $X = [a, b]$ and $Y = [c, d]$ be intervals. Their set difference is $X ∖ Y = \{x: x ∈ X \text{ and } x ∉ Y \}$ and, depending on their position, three different results may occur:

If $X$ and $Y$ do not overlap, i.e., if their intersection is empty, then the set difference is just $X$.

Otherwise, let $Z = X ∩ Y ≠ ∅$, then $Z$ splits $X$ into either one or two intervals. The latter case happens when the bounds of $Y$ are strictly contained in $X$.

To check for strict inclusion, we assume that the inclusion is strict and then check whether the resulting intervals that cover $X$ (one to its left and one to its right, let them be L and R), obtained by intersection with $Y$, are flat. Three cases may arise:

If both L and R are flat then $X = Y$ and the result is the empty set.

If only L is flat, then the result is R, the remaining interval not covered by $Y$. Similarly, if only R is flat, then the result is L.

Finally, if none of the intervals is flat, then $Y$ is strictly contained in $X$ and the set union of L and R is returned.

Examples

julia> X = Interval(0, 2); Y = Interval(1, 4); Z = Interval(2, 3);

julia> difference(X, X)
∅(1)

julia> difference(X, Y)
Interval{Float64}([0, 1])

julia> difference(Y, Z)
UnionSet{Float64, Interval{Float64}, Interval{Float64}}(Interval{Float64}([1, 2]), Interval{Float64}([3, 4]))

Base.isdisjoint — Function

isdisjoint(X::LazySet, Y::LazySet, [witness]::Bool=false)

Check whether two sets are disjoint (i.e., do not intersect), and optionally compute a witness.

Input

X – set

Y – set

witness – (optional, default: false) compute a witness if activated

Output

If the witness option is deactivated: true iff $X ∩ Y = ∅$

If the witness option is activated:

(true, []) iff $X ∩ Y = ∅$

(false, v) iff $X ∩ Y ≠ ∅$ for some $v ∈ X ∩ Y$

Notes

The convenience alias is_intersection_empty is also available.

Base.isdisjoint — Function

Extended help

isdisjoint(I1::Interval, I2::Interval, [witness]::Bool=false)

Algorithm

$I1 ∩ I2 = ∅$ iff there is a gap between the left-most point of the second interval and the right-most point of the first interval, or vice-versa.

A witness is computed by taking the maximum over the left-most points of each interval, which is guaranteed to belong to the intersection.

LazySets.API.minkowski_difference — Method

minkowski_difference(X::LazySet, Y::LazySet)

Compute the Minkowski difference of two sets.

Input

X – set

Y – set

Output

A set representing the Minkowski difference $X ⊖ Y$.

Notes

The Minkowski difference of two sets $X$ and $Y$ is defined as

\[ X ⊖ Y = \{z \mid \{z\} ⊕ Y ⊆ X\}\]

The convenience alias pontryagin_difference is also available.

There is some inconsistency in the literature regarding the naming conventions. In this library, both Minkowski difference and Pontryagin difference refer to the geometric difference of two sets.

LazySets.API.minkowski_difference — Method

Extended help

minkowski_difference(I1::Interval, I2::Interval)

Output

An Interval, or an EmptySet if the difference is empty.

LazySets.plot_recipe — Method

plot_recipe(X::Interval{N}, [ε]=zero(N)) where {N}

Convert an interval to a pair (x, y) of points for plotting.

Input

X – interval

ε – (optional, default: 0) ignored, used for dispatch

Output

A pair (x, y) of two points that can be plotted.

Notes

We consider the interval as a line segment with y coordinate equal to zero.

Base.min — Method

min(X::Interval)

Return the lower component of an interval.

Input

X – interval

Output

The lower (lo) component of the interval (a number).

Base.max — Method

max(X::Interval)

Return the higher or upper component of an interval.

Input

X – interval

Output

The higher (hi) component of the interval (a number).

Undocumented implementations:

center

center

complement

constraints_list

copy(::Type{LazySet})

diameter

dim

extrema

extrema

high

high

isoperationtype

low

low

norm

radius

reflect

volume

affine_map

exponential_map

∈

permute

project

scale

ρ

σ

convex_hull

distance

intersection

isapprox

isequivalent

⊂

⊆

minkowski_sum

Inherited from LazySet:

concretize

constraints

eltype

eltype

isoperation

polyhedron

rationalize

singleton_list

tosimplehrep

vertices

is_interior_point

sample

==

Inherited from LazySet but does not apply:

triangulate

triangulate_faces

Inherited from AbstractPolyhedron:

ispolyhedral

Inherited from ConvexSet:

isconvextype

linear_combination

Inherited from AbstractPolytope:

isbounded

isboundedtype

Inherited from AbstractCentrallySymmetricPolytope:

isempty

isuniversal

Inherited from AbstractZonotope:

order

togrep

Inherited from AbstractHyperrectangle:

generators

genmat

cartesian_product

Some additional functionality is available for IntervalArithmetic.Intervals:

LazySets.fast_interval_pow — Method

fast_interval_pow(a::IntervalArithmetic.Interval, n::Int)

Compute the nth power of an interval without using correct rounding.

Input

a – interval (from IntervalArithmetic.jl)

n – integer

Output

A non-rigorous approximation of a^n.

Notes

For a rigorous approximation with correct rounding, use a^n from IntervalArithmetic.jl.