Rotated hyperrectangle (RotatedHyperrectangle)

LazySets.RotatedHyperrectangle — Type

RotatedHyperrectangle{N, MN<:AbstractMatrix{N},
                      HT<:AbstractHyperrectangle{N}} <: AbstractZonotope{N}

Type that represents a hyperrectangle that is not necessarily axis aligned.

Fields

M – matrix (not necessarily invertible)

box – axis-aligned hyperrectangle

Notes

The matrix M is typically a rotation matrix and hence invertible, but this type does not require that and any matrix that is compatible with the dimension of box is allowed.

LazySets.API.dim — Method

dim(R::RotatedHyperrectangle)

Return the ambient dimension of a rotated hyperrectangle.

Input

R – rotated hyperrectangle

Output

The ambient dimension of the rotated hyperrectangle.

LazySets.API.ρ — Method

ρ(d::AbstractVector, R::RotatedHyperrectangle)

Evaluate the support function of a rotated hyperrectangle in a given direction.

Input

d – direction

R – rotated hyperrectangle

Output

Evaluation of the support function in the given direction.

LazySets.API.σ — Method

σ(d::AbstractVector, R::RotatedHyperrectangle)

Return a support vector of a rotated hyperrectangle in a given direction.

Input

d – direction

R – rotated hyperrectangle

Output

A support vector in the given direction.

LazySets.API.center — Method

center(R::RotatedHyperrectangle)

Return the center of a rotated hyperrectangle.

Input

R – rotated hyperrectangle

Output

The center of the rotated hyperrectangle.

LazySets.generators — Method

generators(R::RotatedHyperrectangle)

Return an iterator over the generators of a rotated hyperrectangle.

Input

R – rotated hyperrectangle

Output

An iterator over the generators of R.

LazySets.genmat — Method

genmat(R::RotatedHyperrectangle)

Return the generator matrix of a rotated hyperrectangle.

Input

R – rotated hyperrectangle

Output

A matrix where each column represents one generator of R.

LazySets.API.linear_map — Method

linear_map(M::AbstractMatrix, R::RotatedHyperrectangle)

Compute the concrete linear map of a rotated hyperrectangle.

Input

M – matrix

R – rotated hyperrectangle

Output

A new rotated hyperrectangle.

Notes

If M is not a square matrix, the result is not necessarily a rotated hyperrectangle. The type still represents the correct set, but it is suggested to only apply square linear maps to it.

LazySets.API.vertices_list — Method

vertices_list(R::RotatedHyperrectangle)

Return the list of vertices of a rotated hyperrectangle.

Input

R – rotated hyperrectangle

Output

A list of the vertices.

LazySets.API.constraints_list — Method

constraints_list(R::RotatedHyperrectangle)

Return the list of constraints of a rotated hyperrectangle.

Input

R – rotated hyperrectangle

Output

A list of the linear constraints.

Inherited from LazySet:

high

low

norm

radius

diameter

singleton_list

area

rectify

Inherited from AbstractPolytope:

isbounded

isuniversal

volume

Inherited from AbstractCentrallySymmetricPolytope:

isempty

an_element

Inherited from AbstractZonotope:

∈

translate

constraints_list

order

reflect