Singleton

LazySets.Singleton — Type

Singleton{N, VN<:AbstractVector{N}} <: AbstractSingleton{N}

Type that represents a singleton, that is, a set with a single element.

Fields

element – the only element of the set

Examples

julia> Singleton([1.0, 2.0])
Singleton{Float64, Vector{Float64}}([1.0, 2.0])

julia> Singleton(1.0, 2.0)  # convenience constructor from numbers
Singleton{Float64, Vector{Float64}}([1.0, 2.0])

Base.rand — Method

rand(::Type{Singleton}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Create a random singleton.

Input

Singleton – type for dispatch

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

A random singleton.

Algorithm

The element is a normally distributed vector with entries of mean 0 and standard deviation 1.

LazySets.element — Method

element(S::Singleton)

Return the element of a singleton.

Input

S – singleton

Output

The element of the singleton.

LazySets.translate — Method

translate(S::Singleton, v::AbstractVector)

Translate (i.e., shift) a singleton by a given vector.

Input

S – singleton

v – translation vector

Output

A translated singleton.

Notes

See also translate!(::Singleton, ::AbstractVector) for the in-place version.

LazySets.translate! — Method

translate!(S::Singleton, v::AbstractVector)

Translate (i.e., shift) a singleton by a given vector, in-place.

Input

S – singleton

v – translation vector

Output

The singleton S translated by v.

Algorithm

We add the vector to the point in the singleton.

Notes

See also translate(::Singleton, ::AbstractVector) for the out-of-place version.

ReachabilityBase.Arrays.rectify — Method

rectify(S::Singleton)

Concrete rectification of a singleton.

Input

S – singleton

Output

The Singleton that corresponds to the rectification of S.

LazySets.project — Method

project(S::Singleton, block::AbstractVector{Int}; kwargs...)

Concrete projection of a singleton.

Input

S – singleton

block – block structure, a vector with the dimensions of interest

Output

A set representing the projection of the singleton S on the dimensions specified by block.

SparseArrays.permute — Method

permute(S::Singleton, p::AbstractVector{Int})

Permute the dimensions according to a permutation vector.

Input

S – singleton

p – permutation vector

Output

A new Singleton with the permuted dimensions.

LazySets.singleton_list — Method

singleton_list(S::Singleton)

Return the vertices of a singleton as a list of singletons.

Input

S – singleton

Output

The list of vertices of S, as Singleton.

LazySets.linear_map — Method

linear_map(M::AbstractMatrix, S::Singleton)

Concrete linear map of a singleton.

Input

M – matrix

S – singleton

Output

A singleton.

Inherited from LazySet:

diameter

Inherited from AbstractPolytope:

isbounded

isuniversal

Inherited from AbstractCentrallySymmetricPolytope:

dim

isempty

an_element

Inherited from AbstractZonotope:

order

togrep

Inherited from AbstractHyperrectangle:

norm

radius

constraints_list

Inherited from AbstractSingleton:

σ

ρ

∈

center

vertices

vertices_list

radius_hyperrectangle

radius_hyperrectangle

generators

genmat

ngens

high

low

reflect