ZeroSet

Origin (ZeroSet)

LazySets.ZeroSet — Type.

ZeroSet{N<:Real} <: AbstractSingleton{N}

Type that represents the zero set, i.e., the set that only contains the origin.

Fields

dim – the ambient dimension of this zero set

source

LazySets.dim — Method.

dim(Z::ZeroSet)

Return the ambient dimension of this zero set.

Input

Z – a zero set, i.e., a set that only contains the origin

Output

The ambient dimension of the zero set.

source

LazySets.σ — Method.

σ(d::AbstractVector{N}, Z::ZeroSet{N}) where {N<:Real}

Return the support vector of a zero set.

Input

d – direction

Z – a zero set, i.e., a set that only contains the origin

Output

The returned value is the origin since it is the only point that belongs to this set.

source

LazySets.ρ — Method.

ρ(d::AbstractVector{N}, Z::ZeroSet{N}) where {N<:Real}

Evaluate the support function of a zero set in a given direction.

Input

d – direction

Z – a zero set, i.e., a set that only contains the origin

Output

0.

source

Base.:∈ — Method.

∈(x::AbstractVector{N}, Z::ZeroSet{N}) where {N<:Real}

Check whether a given point is contained in a zero set.

Input

x – point/vector

Z – zero set

Output

true iff $x ∈ Z$.

Examples

julia> Z = ZeroSet(2);

julia> [1.0, 0.0] ∈ Z
false
julia> [0.0, 0.0] ∈ Z
true

source

Base.rand — Method.

rand(::Type{ZeroSet}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Create a zero set (note that there is nothing to randomize).

Input

ZeroSet – type for dispatch

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

The (only) zero set of the given numeric type and dimension.

source

LazySets.element — Method.

element(S::ZeroSet{N}) where {N<:Real}

Return the element of a zero set.

Input

S – zero set

Output

The element of the zero set, i.e., a zero vector.

source

LazySets.element — Method.

element(S::ZeroSet{N}, ::Int) where {N<:Real}

Return the i-th entry of the element of a zero set.

Input

S – zero set

i – dimension

Output

The i-th entry of the element of the zero set, i.e., 0.

source

LazySets.linear_map — Method.

linear_map(M::AbstractMatrix{N}, Z::ZeroSet{N}) where {N<:Real}

Concrete linear map of a zero set.

Input

M – matrix

Z – zero set

Output

The zero set whose dimension matches the output dimension of the given matrix.

source

LazySets.translate — Method.

translate(Z::ZeroSet{N}, v::AbstractVector{N}) where {N<:Real}

Translate (i.e., shift) a zero set by a given vector.

Input

Z – zero set

v – translation vector

Output

A singleton containing the vector v.

source

Inherited from LazySet:

diameter

Inherited from AbstractPolytope:

isbounded

isuniversal

singleton_list

Inherited from AbstractCentrallySymmetricPolytope:

isempty

Inherited from AbstractZonotope:

ngens

order

Inherited from AbstractHyperrectangle:

norm

radius

high

low

constraints_list

Inherited from AbstractSingleton:

radius_hyperrectangle

radius_hyperrectangle

vertices_list

center

an_element

generators

genmat