Origin (ZeroSet)

LazySets.ZeroSetModule.ZeroSet — Type

ZeroSet{N} <: AbstractSingleton{N}

Type that represents the zero set, i.e., the set that only contains the origin.

Fields

dim – the ambient dimension of the set

Operations

LazySets.element — Method

element(Z::ZeroSet)

Return the element of a zero set.

Input

Z – zero set

Output

The element of the zero set, i.e., a zero vector.

LazySets.element — Method

element(Z::ZeroSet, i::Int)

Return the i-th entry of the element of a zero set.

Input

Z – zero set

i – dimension

Output

The i-th entry of the element of the zero set, i.e., 0.

Base.rand — Method

rand(T::Type{<:LazySet}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing
    )

Create a random set of the given set type.

Input

T – set type

N – (optional, default: Float64) numeric type

dim – (optional, default: 2) dimension

rng – (optional, default: GLOBAL_RNG) random number generator

seed – (optional, default: nothing) seed for reseeding

Output

A random set of the given set type.

Base.rand — Method

Extended help

rand(::Type{ZeroSet}; [N]::Type{<:Real}=Float64, [dim]::Int=2,
     [rng]::AbstractRNG=GLOBAL_RNG, [seed]::Union{Int, Nothing}=nothing)

Output

The (only) zero set of the given numeric type and dimension.

Undocumented implementations:

copy(::Type{ZeroSet})

dim

reflect

rectify

∈

isoperationtype

linear_map

scale

scale!

ρ

translate

minkowski_sum

Inherited from LazySet:

complement

concretize

constraints

convex_hull

diameter

eltype

eltype

isoperation

polyhedron

rationalize

singleton_list

tosimplehrep

triangulate

triangulate_faces

affine_map

exponential_map

is_interior_point

sample

convex_hull

≈

==

⊂

Inherited from ConvexSet:

isconvextype

linear_combination

Inherited from AbstractPolyhedron:

ispolyhedral

Inherited from AbstractPolytope:

isbounded

isboundedtype

Inherited from AbstractCentrallySymmetricPolytope:

an_element

isempty

isuniversal

Inherited from AbstractZonotope:

order

remove_redundant_generators

togrep

reduce_order

split

split

Inherited from AbstractHyperrectangle:

area

extrema

extrema

isflat

norm

radius

volume

project

difference

Inherited from AbstractSingleton:

center

center

chebyshev_center_radius

constraints_list

generators

genmat

high

high

low

low

ngens

radius_hyperrectangle

radius_hyperrectangle

vertices

vertices

vertices_list

σ

cartesian_product

distance

intersection

isdisjoint

isequivalent

⊆